@@ -7,9 +7,11 @@ import 'package:html/dom.dart' as dom;
 
 import 'core_html_widget.dart';
 
-export 'widgets/css_block.dart';
+export 'external/csslib.dart';
 export 'widgets/css_sizing.dart';
+export 'widgets/html_list_item.dart';
 export 'widgets/html_ruby.dart';
+export 'widgets/html_table.dart';
 
 /// The default character threshold to build widget tree asynchronously.
 ///
@@ -32,7 +34,8 @@ const widget0 = SizedBox.shrink();
 ///     element.classes.contains('name') ? {'color': 'red'} : null,
 /// )
 /// ```
-typedef CustomStylesBuilder = Map<String, String> Function(dom.Element element);
+typedef CustomStylesBuilder = Map<String, String>? Function(
+    dom.Element element);
 
 /// A callback to render custom widget for a DOM element.
 ///
@@ -40,7 +43,7 @@ typedef CustomStylesBuilder = Map<String, String> Function(dom.Element element);
 /// you have to handle the DOM element and its children manually,
 /// if the children have HTML styling etc., they won't be processed at all.
 /// For those needs, a custom [WidgetFactory] is the way to go.
-typedef CustomWidgetBuilder = Widget Function(dom.Element element);
+typedef CustomWidgetBuilder = Widget? Function(dom.Element element);
 
 /// A set of values that should trigger rebuild.
 class RebuildTriggers {
@@ -78,17 +81,17 @@ class WidgetPlaceholder<T> extends StatelessWidget {
   /// The origin of this widget.
   final T generator;
 
-  final List<Widget Function(BuildContext, Widget)> _builders = [];
-  final Widget _firstChild;
+  final List<Widget? Function(BuildContext, Widget)> _builders = [];
+  final Widget? _firstChild;
 
   /// Creates a widget builder.
-  WidgetPlaceholder(this.generator, {Widget child}) : _firstChild = child;
+  WidgetPlaceholder(this.generator, {Widget? child}) : _firstChild = child;
 
   @override
   Widget build(BuildContext context) => callBuilders(context, _firstChild);
 
   /// Calls builder callbacks on the specified [child] widget.
-  Widget callBuilders(BuildContext context, Widget child) {
+  Widget callBuilders(BuildContext context, Widget? child) {
     var built = child ?? widget0;
 
     for (final builder in _builders) {
@@ -110,8 +113,7 @@ class WidgetPlaceholder<T> extends StatelessWidget {
 
   /// Enqueues [builder] to be built later.
   WidgetPlaceholder<T> wrapWith(
-      Widget Function(BuildContext context, Widget child) builder) {
-    assert(builder != null);
+      Widget? Function(BuildContext context, Widget child) builder) {
     _builders.add(builder);
     return this;
   }
@@ -130,7 +132,7 @@ class WidgetPlaceholder<T> extends StatelessWidget {
       : WidgetPlaceholder<Widget>(child, child: child);
 }
 
-final _dataUriRegExp = RegExp(r'^data:image/[^;]+;(base64|utf8),');
+final _dataUriRegExp = RegExp(r'^data:[^;]+;([^,]+),');
 
 /// Returns [Uint8List] by decoding [dataUri].
 ///
@@ -138,29 +140,29 @@ final _dataUriRegExp = RegExp(r'^data:image/[^;]+;(base64|utf8),');
 ///
 /// - base64
 /// - utf8
-Uint8List bytesFromDataUri(String dataUri) {
+Uint8List? bytesFromDataUri(String dataUri) {
   final match = _dataUriRegExp.matchAsPrefix(dataUri);
   if (match == null) return null;
 
-  final prefix = match[0];
+  final prefix = match[0]!;
   final encoding = match[1];
   final data = dataUri.substring(prefix.length);
-
   final bytes = encoding == 'base64'
       ? base64.decode(data)
-      : encoding == 'utf8' ? Uint8List.fromList(data.codeUnits) : null;
-  if (bytes.isEmpty) return null;
+      : encoding == 'utf8'
+          ? Uint8List.fromList(data.codeUnits)
+          : null;
 
-  return bytes;
+  return bytes?.isNotEmpty == true ? bytes : null;
 }
 
 /// Returns [List<T>] if [x] is provided or `null` otherwise.
-Iterable<T> listOrNull<T>(T x) => x == null ? null : [x];
+Iterable<T>? listOrNull<T>(T? x) => x == null ? null : [x];
 
 /// Parses [key] from [map] as an double literal and return its value.
-double tryParseDoubleFromMap(Map<dynamic, String> map, String key) =>
-    map.containsKey(key) ? double.tryParse(map[key]) : null;
+double? tryParseDoubleFromMap(Map<dynamic, String> map, String key) =>
+    map.containsKey(key) ? double.tryParse(map[key]!) : null;
 
 /// Parses [key] from [map] as a, possibly signed, integer literal and return its value.
-int tryParseIntFromMap(Map<dynamic, String> map, String key) =>
-    map.containsKey(key) ? int.tryParse(map[key]) : null;
+int? tryParseIntFromMap(Map<dynamic, String> map, String key) =>
+    map.containsKey(key) ? int.tryParse(map[key]!) : null;
