@@ -5,66 +5,61 @@ part of '../core_data.dart';
 class TextStyleHtml {
   final Iterable<dynamic> _deps;
 
-  /// The cross axis alignment.
-  ///
-  /// Default: [CrossAxisAlignment.start].
-  final CrossAxisAlignment crossAxisAlignment;
-
   /// The line height.
-  final double height;
+  final double? height;
 
   /// The number of max lines that should be rendered.
-  final int maxLines;
+  final int? maxLines;
 
   /// The parent style.
-  final TextStyleHtml parent;
+  final TextStyleHtml? parent;
 
   /// The input [TextStyle].
   final TextStyle style;
 
   /// The text alignment.
-  final TextAlign textAlign;
+  final TextAlign? textAlign;
 
   /// The text direction.
   final TextDirection textDirection;
 
   /// The overflow behavior.
-  final TextOverflow textOverflow;
+  final TextOverflow? textOverflow;
 
-  /// Creates a text style.
-  TextStyleHtml({
-    this.crossAxisAlignment,
-    @required Iterable deps,
+  /// The whitespace behavior.
+  final CssWhitespace whitespace;
+
+  TextStyleHtml._({
+    required Iterable<dynamic> deps,
     this.height,
     this.maxLines,
     this.parent,
-    this.style,
+    required this.style,
     this.textAlign,
-    this.textDirection,
+    required this.textDirection,
     this.textOverflow,
+    required this.whitespace,
   }) : _deps = deps;
 
   /// Creates the root text style.
-  factory TextStyleHtml.root(
-      Iterable<dynamic> deps, TextStyle widgetTextStyle) {
+  factory TextStyleHtml.root(Iterable<dynamic> deps, TextStyle? widgetStyle) {
     var style = _getDependency<TextStyle>(deps);
-    if (widgetTextStyle != null) {
-      style = widgetTextStyle.inherit
-          ? style.merge(widgetTextStyle)
-          : widgetTextStyle;
+    if (widgetStyle != null) {
+      style = widgetStyle.inherit ? style.merge(widgetStyle) : widgetStyle;
     }
 
-    var mqd = _getDependency<MediaQueryData>(deps);
+    final mqd = _getDependency<MediaQueryData>(deps);
     final tsf = mqd.textScaleFactor;
-    if (tsf != 1) {
-      style = style.copyWith(fontSize: style.fontSize * tsf);
+    final fontSize = style.fontSize;
+    if (tsf != 1 && fontSize != null) {
+      style = style.copyWith(fontSize: fontSize * tsf);
     }
 
-    return TextStyleHtml(
+    return TextStyleHtml._(
       deps: deps,
-      parent: null,
       style: style,
       textDirection: _getDependency<TextDirection>(deps),
+      whitespace: CssWhitespace.normal,
     );
   }
 
@@ -74,21 +69,20 @@ class TextStyleHtml {
   /// `TextStyle` with existing height cannot be copied with `height=null`.
   /// See [flutter/flutter#58765](https://github.com/flutter/flutter/issues/58765).
   TextStyle get styleWithHeight =>
-      height != null && height >= 0 ? style.copyWith(height: height) : style;
+      height != null && height! >= 0 ? style.copyWith(height: height) : style;
 
   /// Creates a copy with the given fields replaced with the new values.
   TextStyleHtml copyWith({
-    CrossAxisAlignment crossAxisAlignment,
-    double height,
-    int maxLines,
-    TextStyleHtml parent,
-    TextStyle style,
-    TextAlign textAlign,
-    TextDirection textDirection,
-    TextOverflow textOverflow,
+    double? height,
+    int? maxLines,
+    TextStyleHtml? parent,
+    TextStyle? style,
+    TextAlign? textAlign,
+    TextDirection? textDirection,
+    TextOverflow? textOverflow,
+    CssWhitespace? whitespace,
   }) =>
-      TextStyleHtml(
-        crossAxisAlignment: crossAxisAlignment ?? this.crossAxisAlignment,
+      TextStyleHtml._(
         deps: _deps,
         height: height ?? this.height,
         maxLines: maxLines ?? this.maxLines,
@@ -97,6 +91,7 @@ class TextStyleHtml {
         textAlign: textAlign ?? this.textAlign,
         textDirection: textDirection ?? this.textDirection,
         textOverflow: textOverflow ?? this.textOverflow,
+        whitespace: whitespace ?? this.whitespace,
       );
 
   /// Gets dependency value by type.
@@ -109,19 +104,19 @@ class TextStyleHtml {
       return value;
     }
 
-    return null;
+    throw StateError('The $T dependency could not be found');
   }
 }
 
 /// A text styling builder.
 class TextStyleBuilder<T1> {
   /// The parent builder.
-  final TextStyleBuilder parent;
+  final TextStyleBuilder? parent;
 
-  List<Function> _builders;
-  List _inputs;
-  TextStyleHtml _parentOutput;
-  TextStyleHtml _output;
+  List<Function>? _builders;
+  List? _inputs;
+  TextStyleHtml? _parentOutput;
+  TextStyleHtml? _output;
 
   /// Create a builder.
   TextStyleBuilder({this.parent});
@@ -129,16 +124,14 @@ class TextStyleBuilder<T1> {
   /// Enqueues a callback.
   void enqueue<T2>(
     TextStyleHtml Function(TextStyleHtml tsh, T2 input) builder, [
-    T2 input,
+    T2? input,
   ]) {
-    if (builder == null) return;
-
     assert(_output == null, 'Cannot add builder after being built');
     _builders ??= [];
-    _builders.add(builder);
+    _builders!.add(builder);
 
     _inputs ??= [];
-    _inputs.add(input);
+    _inputs!.add(input);
   }
 
   /// Builds a [TextStyleHtml] by calling queued callbacks.
@@ -149,34 +142,41 @@ class TextStyleBuilder<T1> {
       _output = null;
     }
 
-    if (_output != null) return _output;
-    if (_builders == null) return _output = _parentOutput;
+    if (_output != null) return _output!;
+    if (_builders == null) return _output = _parentOutput!;
 
     _output = _parentOutput?.copyWith(parent: _parentOutput);
-    final l = _builders.length;
+    final l = _builders!.length;
     for (var i = 0; i < l; i++) {
-      final builder = _builders[i];
-      _output = builder(_output, _inputs[i]);
+      final builder = _builders![i];
+      _output = builder(_output, _inputs![i]);
       assert(_output?.parent == _parentOutput);
     }
 
-    return _output;
+    return _output!;
   }
 
   /// Returns `true` if this shares same styling with [other].
-  bool hasSameStyleWith(TextStyleBuilder other) {
+  bool hasSameStyleWith(TextStyleBuilder? other) {
     if (other == null) return false;
-
-    var thisWithBuilder = this;
+    TextStyleBuilder thisWithBuilder = this;
     while (thisWithBuilder._builders == null) {
-      if (thisWithBuilder.parent == null) break;
-      thisWithBuilder = thisWithBuilder.parent;
+      final thisParent = thisWithBuilder.parent;
+      if (thisParent == null) {
+        break;
+      } else {
+        thisWithBuilder = thisParent;
+      }
     }
 
     var otherWithBuilder = other;
     while (otherWithBuilder._builders == null) {
-      if (otherWithBuilder.parent == null) break;
-      otherWithBuilder = otherWithBuilder.parent;
+      final otherParent = otherWithBuilder.parent;
+      if (otherParent == null) {
+        break;
+      } else {
+        otherWithBuilder = otherParent;
+      }
     }
 
     return thisWithBuilder == otherWithBuilder;
