@@ -1,12 +1,15 @@
-import 'package:flutter/widgets.dart';
+import 'dart:collection';
+
+import 'package:csslib/parser.dart' as css;
+import 'package:csslib/visitor.dart' as css;
+
 import 'package:html/dom.dart' as dom;
 
 import '../core_data.dart';
 import '../core_data.dart' as core_data show BuildMetadata, BuildTree;
 import '../core_helpers.dart';
 import '../core_widget_factory.dart';
-import 'core_parser.dart';
-import 'flattener.dart';
+import 'core_ops.dart';
 
 final _regExpSpaceLeading = RegExp(r'^[^\S\u{00A0}]+', unicode: true);
 final _regExpSpaceTrailing = RegExp(r'[^\S\u{00A0}]+$', unicode: true);
@@ -15,83 +18,51 @@ final _regExpSpaces = RegExp(r'[^\S\u{00A0}]+', unicode: true);
 class BuildMetadata extends core_data.BuildMetadata {
   final Iterable<BuildOp> _parentOps;
 
-  List<BuildOp> _buildOps;
+  Set<BuildOp>? _buildOps;
   var _buildOpsIsLocked = false;
-  bool _isBlockElement;
-  List<String> _styles;
+  List<css.Declaration>? _styles;
   var _stylesIsLocked = false;
+  bool? _willBuildSubtree;
 
-  BuildMetadata(dom.Element element, TextStyleBuilder tsb, [this._parentOps])
+  BuildMetadata(dom.Element element, TextStyleBuilder tsb,
+      [this._parentOps = const []])
       : super(element, tsb);
 
   @override
-  Iterable<BuildOp> get buildOps => _buildOps;
-
-  @override
-  bool get isBlockElement => _isBlockElement ?? _isBlockElementFrom(_buildOps);
+  Iterable<BuildOp> get buildOps => _buildOps ?? const [];
 
   @override
   Iterable<BuildOp> get parentOps => _parentOps;
 
   @override
-  Iterable<MapEntry<String, String>> get styles sync* {
+  List<css.Declaration> get styles {
     assert(_stylesIsLocked);
-
-    final iterator = _styles?.iterator;
-    while (iterator?.moveNext() == true) {
-      final key = iterator.current;
-      if (!iterator.moveNext()) return;
-      yield MapEntry(key, iterator.current);
-    }
+    return _styles ?? const [];
   }
 
   @override
-  set isBlockElement(bool value) {
-    assert(!_buildOpsIsLocked, 'Metadata can no longer be changed.');
-    _isBlockElement = value;
-  }
+  bool? get willBuildSubtree => _willBuildSubtree;
 
   @override
   operator []=(String key, String value) {
-    if (key == null || value == null) return;
-
     assert(!_stylesIsLocked, 'Metadata can no longer be changed.');
+    final styleSheet = css.parse('*{$key: $value;}');
     _styles ??= [];
-    _styles..add(key)..add(value);
+    _styles!.addAll(styleSheet.collectDeclarations());
   }
 
   @override
   void register(BuildOp op) {
-    if (op == null) return;
-
     assert(!_buildOpsIsLocked, 'Metadata can no longer be changed.');
-    _buildOps ??= [];
-    if (!buildOps.contains(op)) _buildOps.add(op);
+    _buildOps ??= SplayTreeSet(_compareBuildOps);
+    _buildOps!.add(op);
   }
-
-  void _sortBuildOps() {
-    assert(!_buildOpsIsLocked);
-
-    if (_buildOps != null) {
-      _buildOps.sort((a, b) => a.priority.compareTo(b.priority));
-
-      // pre-calculate `isBlockElement` for faster access later
-      _isBlockElement ??= _isBlockElementFrom(_buildOps);
-    } else {
-      _isBlockElement ??= false;
-    }
-
-    _buildOpsIsLocked = true;
-  }
-
-  static bool _isBlockElementFrom(Iterable<BuildOp> ops) =>
-      ops?.where((op) => op.isBlockElement)?.length?.compareTo(0) == 1;
 }
 
 class BuildTree extends core_data.BuildTree {
-  final CustomStylesBuilder customStylesBuilder;
-  final CustomWidgetBuilder customWidgetBuilder;
-  final BuildMetadata parentMeta;
+  final CustomStylesBuilder? customStylesBuilder;
+  final CustomWidgetBuilder? customWidgetBuilder;
+  final core_data.BuildMetadata parentMeta;
   final Iterable<BuildOp> parentOps;
   final WidgetFactory wf;
 
@@ -100,15 +71,15 @@ class BuildTree extends core_data.BuildTree {
   BuildTree({
     this.customStylesBuilder,
     this.customWidgetBuilder,
-    BuildTree parent,
-    this.parentMeta,
-    this.parentOps,
-    TextStyleBuilder tsb,
-    this.wf,
+    core_data.BuildTree? parent,
+    required this.parentMeta,
+    this.parentOps = const [],
+    required TextStyleBuilder tsb,
+    required this.wf,
   }) : super(parent, tsb);
 
   @override
-  BuildBit add(BuildBit bit) {
+  T add<T extends BuildBit>(T bit) {
     assert(_built.isEmpty, "Built tree shouldn't be altered.");
     return super.add(bit);
   }
@@ -117,11 +88,8 @@ class BuildTree extends core_data.BuildTree {
     for (final domNode in domNodes) {
       _addBitsFromNode(domNode);
     }
-
-    if (parentMeta?.buildOps != null) {
-      for (final op in parentMeta.buildOps) {
-        op.onTree?.call(parentMeta, this);
-      }
+    for (final op in parentMeta.buildOps) {
+      op.onTree?.call(parentMeta, this);
     }
   }
 
@@ -129,16 +97,13 @@ class BuildTree extends core_data.BuildTree {
   Iterable<WidgetPlaceholder> build() {
     if (_built.isNotEmpty) return _built;
 
-    var widgets = _flatten();
-
-    if (parentMeta?.buildOps != null) {
-      for (final op in parentMeta.buildOps) {
-        widgets = op.onWidgets
-                ?.call(parentMeta, widgets)
-                ?.map(WidgetPlaceholder.lazy)
-                ?.toList(growable: false) ??
-            widgets;
-      }
+    var widgets = wf.flatten(parentMeta, this);
+    for (final op in parentMeta.buildOps) {
+      widgets = op.onWidgets
+              ?.call(parentMeta, widgets)
+              ?.map(WidgetPlaceholder.lazy)
+              .toList(growable: false) ??
+          widgets;
     }
 
     _built.addAll(widgets);
@@ -147,10 +112,10 @@ class BuildTree extends core_data.BuildTree {
 
   @override
   BuildTree sub({
-    core_data.BuildTree parent,
-    BuildMetadata parentMeta,
-    Iterable<BuildOp> parentOps,
-    TextStyleBuilder tsb,
+    core_data.BuildTree? parent,
+    BuildMetadata? parentMeta,
+    Iterable<BuildOp> parentOps = const [],
+    TextStyleBuilder? tsb,
   }) =>
       BuildTree(
         customStylesBuilder: customStylesBuilder,
@@ -163,31 +128,33 @@ class BuildTree extends core_data.BuildTree {
       );
 
   void _addBitsFromNode(dom.Node domNode) {
-    if (domNode.nodeType == dom.Node.TEXT_NODE) return _addText(domNode.text);
+    if (domNode.nodeType == dom.Node.TEXT_NODE) {
+      final text = domNode as dom.Text;
+      return _addText(text.data);
+    }
     if (domNode.nodeType != dom.Node.ELEMENT_NODE) return;
 
     final element = domNode as dom.Element;
-    final meta = BuildMetadata(element, parentMeta.tsb().sub(), parentOps);
     final customWidget = customWidgetBuilder?.call(element);
-    if (customWidget != null) meta.isBlockElement = true;
+    if (customWidget != null) {
+      add(WidgetBit.block(this, customWidget));
+      // skip further processing if a custom widget found
+      return;
+    }
 
+    final meta = BuildMetadata(element, parentMeta.tsb.sub(), parentOps);
     _collectMetadata(meta);
 
     final subTree = sub(
       parentMeta: meta,
       parentOps: _prepareParentOps(parentOps, meta),
-      tsb: meta.tsb(),
+      tsb: meta.tsb,
     );
     add(subTree);
 
-    if (customWidget != null) {
-      // skip sub tree processing if a custom widget found
-      subTree.add(WidgetBit.block(subTree, customWidget));
-    } else {
-      subTree.addBitsFromNodes(element.nodes);
-    }
+    subTree.addBitsFromNodes(element.nodes);
 
-    if (meta.isBlockElement) {
+    if (meta.willBuildSubtree == true) {
       for (final widget in subTree.build()) {
         add(WidgetBit.block(this, widget));
       }
@@ -202,62 +169,79 @@ class BuildTree extends core_data.BuildTree {
     final end = trailing == null ? data.length : trailing.start;
 
     if (end <= start) {
-      addWhitespace();
+      // the string contains all spaces
+      addWhitespace(data);
       return;
     }
 
-    if (start > 0) addWhitespace();
+    if (start > 0) addWhitespace(leading!.group(0)!);
 
-    final substring = data.substring(start, end);
-    final dedup = substring.replaceAll(_regExpSpaces, ' ');
-    addText(dedup);
+    final contents = data.substring(start, end);
+    final spaces = _regExpSpaces.allMatches(contents);
+    var offset = 0;
+    for (final space in [...spaces, null]) {
+      if (space == null) {
+        // reaches end of string
+        final text = contents.substring(offset);
+        if (text.isNotEmpty) {
+          addText(text);
+        }
+        break;
+      } else {
+        final spaceData = space.group(0)!;
+        if (spaceData == ' ') {
+          // micro optimization: ignore single space (ASCII 32)
+          continue;
+        }
 
-    if (end < data.length) addWhitespace();
+        final text = contents.substring(offset, space.start);
+        addText(text);
+
+        addWhitespace(spaceData);
+        offset = space.end;
+      }
+    }
+
+    if (end < data.length) addWhitespace(trailing!.group(0)!);
   }
 
   void _collectMetadata(BuildMetadata meta) {
     wf.parse(meta);
 
-    if (meta.parentOps != null) {
-      for (final op in meta.parentOps) {
-        op.onChild?.call(meta);
-      }
+    for (final op in meta.parentOps) {
+      op.onChild?.call(meta);
     }
 
     // stylings, step 1: get default styles from tag-based build ops
-    if (meta.buildOps != null) {
-      for (final op in meta.buildOps) {
-        final map = op.defaultStyles?.call(meta.element);
-        if (map == null) continue;
+    for (final op in meta.buildOps) {
+      final map = op.defaultStyles?.call(meta.element);
+      if (map == null) continue;
 
-        meta._styles ??= [];
-        for (final pair in map.entries) {
-          if (pair.key == null || pair.value == null) continue;
-          meta._styles.insertAll(0, [pair.key, pair.value]);
-        }
-      }
+      final str = map.entries.map((e) => '${e.key}: ${e.value}').join(';');
+      final styleSheet = css.parse('*{$str}');
+
+      meta._styles ??= [];
+      meta._styles!.insertAll(0, styleSheet.collectDeclarations());
     }
 
     _customStylesBuilder(meta);
 
     // stylings, step 2: get styles from `style` attribute
-    final attrs = meta.element.attributes;
-    if (attrs.containsKey('style')) {
-      for (final pair in splitAttributeStyle(attrs['style'])) {
-        meta[pair.key] = pair.value;
-      }
+    for (final declaration in meta.element.styles) {
+      meta._styles ??= [];
+      meta._styles!.add(declaration);
     }
 
     meta._stylesIsLocked = true;
     for (final style in meta.styles) {
-      wf.parseStyle(meta, style.key, style.value);
+      wf.parseStyle(meta, style);
     }
 
-    if (meta.isBlockElement) {
-      meta.register(wf.styleDisplayBlock());
-    }
+    wf.parseStyleDisplay(meta, meta[kCssDisplay]?.term);
 
-    meta._sortBuildOps();
+    meta._willBuildSubtree = meta[kCssDisplay]?.term == kCssDisplayBlock ||
+        meta._buildOps?.where(_opRequiresBuildingSubtree).isNotEmpty == true;
+    meta._buildOpsIsLocked = true;
   }
 
   void _customStylesBuilder(BuildMetadata meta) {
@@ -268,49 +252,29 @@ class BuildTree extends core_data.BuildTree {
       meta[pair.key] = pair.value;
     }
   }
+}
 
-  Iterable<WidgetPlaceholder> _flatten() {
-    final widgets = <WidgetPlaceholder>[];
+int _compareBuildOps(BuildOp a, BuildOp b) {
+  if (identical(a, b)) return 0;
 
-    for (final flattened in flatten(this)) {
-      if (flattened.widget != null) {
-        widgets.add(WidgetPlaceholder.lazy(flattened.widget));
-        continue;
-      }
-
-      if (flattened.widgetBuilder != null) {
-        widgets.add(WidgetPlaceholder<BuildTree>(this)
-            .wrapWith((context, _) => flattened.widgetBuilder(context)));
-        continue;
-      }
-
-      if (flattened.spanBuilder == null) continue;
-      widgets.add(WidgetPlaceholder<BuildTree>(this).wrapWith((context, _) {
-        final span = flattened.spanBuilder(context);
-        if (span == null || span is! InlineSpan) return widget0;
-
-        final tsh = tsb?.build(context);
-        final textAlign = tsh?.textAlign ?? TextAlign.start;
-
-        if (span is WidgetSpan &&
-            span.alignment == PlaceholderAlignment.baseline &&
-            textAlign == TextAlign.start) {
-          return span.child;
-        }
-
-        return wf.buildText(parentMeta, tsh, span);
-      }));
-    }
-
-    return widgets;
+  final cmp = a.priority.compareTo(b.priority);
+  if (cmp == 0) {
+    // if two ops have the same priority, they should not be considered equal
+    // fallback to compare hash codes for stable sorting
+    // while still provide pseudo random order across different runs
+    return a.hashCode.compareTo(b.hashCode);
+  } else {
+    return cmp;
   }
 }
 
+bool _opRequiresBuildingSubtree(BuildOp op) =>
+    op.onWidgets != null && !op.onWidgetsIsOptional;
+
 Iterable<BuildOp> _prepareParentOps(Iterable<BuildOp> ops, BuildMetadata meta) {
   // try to reuse existing list if possible
-  final withOnChild =
-      meta.buildOps?.where((op) => op.onChild != null)?.toList();
-  return withOnChild?.isNotEmpty != true
+  final withOnChild = meta.buildOps.where((op) => op.onChild != null).toList();
+  return withOnChild.isEmpty
       ? ops
-      : List.unmodifiable([if (ops != null) ...ops, ...withOnChild]);
+      : List.unmodifiable([...ops, ...withOnChild]);
 }
