@@ -9,39 +9,39 @@ class StyleBgColor {
   StyleBgColor(this.wf);
 
   BuildOp get buildOp => BuildOp(
-        isBlockElement: false,
         onTree: (meta, tree) {
-          if (meta.isBlockElement) return;
+          if (meta.willBuildSubtree == true) return;
 
           final bgColor = _parseColor(wf, meta);
           if (bgColor == null) return;
 
           for (final bit in tree.bits) {
-            bit.tsb?.enqueue(_tsb, bgColor);
+            bit.tsb.enqueue(_tsb, bgColor);
           }
         },
         onWidgets: (meta, widgets) {
+          if (meta.willBuildSubtree == false) return widgets;
+
           final color = _parseColor(wf, meta);
           if (color == null) return null;
           return listOrNull(wf.buildColumnPlaceholder(meta, widgets)?.wrapWith(
               (_, child) => wf.buildDecoratedBox(meta, child, color: color)));
         },
-        priority: 15000,
+        onWidgetsIsOptional: true,
+        priority: 6100,
       );
 
-  Color _parseColor(WidgetFactory wf, BuildMetadata meta) {
-    Color color;
+  Color? _parseColor(WidgetFactory wf, BuildMetadata meta) {
+    Color? color;
     for (final style in meta.styles) {
-      switch (style.key) {
-        case kCssBackgroundColor:
-          final parsed = tryParseColor(style.value);
-          if (parsed != null) color = parsed;
-          break;
+      switch (style.property) {
         case kCssBackground:
-          for (final v in splitCssValues(style.value)) {
-            final parsed = tryParseColor(v);
-            if (parsed != null) color = parsed;
+          for (final expression in style.values) {
+            color = tryParseColor(expression) ?? color;
           }
+          break;
+        case kCssBackgroundColor:
+          color = tryParseColor(style.value) ?? color;
           break;
       }
     }
