@@ -33,15 +33,6 @@ const kCssFontStyleNormal = 'normal';
 
 const kCssFontWeight = 'font-weight';
 const kCssFontWeightBold = 'bold';
-const kCssFontWeight100 = '100';
-const kCssFontWeight200 = '200';
-const kCssFontWeight300 = '300';
-const kCssFontWeight400 = '400';
-const kCssFontWeight500 = '500';
-const kCssFontWeight600 = '600';
-const kCssFontWeight700 = '700';
-const kCssFontWeight800 = '800';
-const kCssFontWeight900 = '900';
 
 const kCssLineHeight = 'line-height';
 const kCssLineHeightNormal = 'normal';
@@ -64,10 +55,16 @@ class TextStyleOps {
         ),
       );
 
-  static TextStyleHtml Function(TextStyleHtml, String) fontSize(
-          WidgetFactory wf) =>
-      (p, v) => p.copyWith(
-          style: p.style.copyWith(fontSize: _fontSizeTryParse(wf, p, v)));
+  static TextStyleHtml fontSize(TextStyleHtml p, css.Expression v) =>
+      p.copyWith(style: p.style.copyWith(fontSize: _fontSizeTryParse(p, v)));
+
+  static TextStyleHtml fontSizeEm(TextStyleHtml p, double v) => p.copyWith(
+      style: p.style.copyWith(
+          fontSize:
+              _fontSizeTryParseCssLength(p, CssLength(v, CssLengthUnit.em))));
+
+  static TextStyleHtml fontSizeTerm(TextStyleHtml p, String v) => p.copyWith(
+      style: p.style.copyWith(fontSize: _fontSizeTryParseTerm(p, v)));
 
   static TextStyleHtml fontStyle(TextStyleHtml p, FontStyle fontStyle) =>
       p.copyWith(style: p.style.copyWith(fontStyle: fontStyle));
@@ -75,12 +72,27 @@ class TextStyleOps {
   static TextStyleHtml fontWeight(TextStyleHtml p, FontWeight v) =>
       p.copyWith(style: p.style.copyWith(fontWeight: v));
 
-  static TextStyleHtml Function(TextStyleHtml, String) lineHeight(
+  static TextStyleHtml Function(TextStyleHtml, css.Expression) lineHeight(
           WidgetFactory wf) =>
       (p, v) => p.copyWith(height: _lineHeightTryParse(wf, p, v));
 
   static TextStyleHtml maxLines(TextStyleHtml p, int v) =>
       p.copyWith(maxLines: v);
+
+  static int? maxLinesTryParse(css.Expression? expression) {
+    if (expression is css.LiteralTerm) {
+      if (expression is css.NumberTerm) {
+        return expression.number.ceil();
+      }
+
+      switch (expression.valueAsString) {
+        case kCssMaxLinesNone:
+          return -1;
+      }
+    }
+
+    return null;
+  }
 
   static TextStyleHtml textDeco(TextStyleHtml p, TextDeco v) {
     final pd = p.style.decoration;
@@ -109,39 +121,45 @@ class TextStyleOps {
     );
   }
 
-  static BuildOp textDecoOp(TextDeco v) => BuildOp(
-        onTree: (meta, _) {
-          if (!meta.isBlockElement) meta.tsb<TextDeco>(textDeco, v);
-        },
-      );
-
   static TextStyleHtml textDirection(TextStyleHtml p, String v) {
-    final textDirection = (v == kCssDirectionRtl)
-        ? TextDirection.rtl
-        : v == kCssDirectionLtr ? TextDirection.ltr : null;
-    if (textDirection == null) return p;
+    switch (v) {
+      case kCssDirectionLtr:
+        return p.copyWith(textDirection: TextDirection.ltr);
+      case kCssDirectionRtl:
+        return p.copyWith(textDirection: TextDirection.rtl);
+    }
 
-    return p.copyWith(textDirection: textDirection);
+    return p;
   }
 
   static TextStyleHtml textOverflow(TextStyleHtml p, TextOverflow v) =>
       p.copyWith(textOverflow: v);
 
-  static List<String> fontFamilyTryParse(String value) {
-    final parts = value.split(',');
+  static TextOverflow? textOverflowTryParse(String value) {
+    switch (value) {
+      case kCssTextOverflowClip:
+        return TextOverflow.clip;
+      case kCssTextOverflowEllipsis:
+        return TextOverflow.ellipsis;
+    }
+
+    return null;
+  }
+
+  static List<String> fontFamilyTryParse(List<css.Expression> expressions) {
     final list = <String>[];
 
-    for (final part in parts) {
-      final fontFamily = part
-          .trim()
-          .replaceFirstMapped(RegExp(r"""^("|')(.+)\1$"""), (m) => m.group(2));
-      if (fontFamily.isNotEmpty) list.add(fontFamily);
+    for (final expression in expressions) {
+      if (expression is css.LiteralTerm) {
+        final fontFamily = expression.valueAsString;
+        if (fontFamily.isNotEmpty) list.add(fontFamily);
+      }
     }
 
     return list;
   }
 
-  static FontStyle fontStyleTryParse(String value) {
+  static FontStyle? fontStyleTryParse(String value) {
     switch (value) {
       case kCssFontStyleItalic:
         return FontStyle.italic;
@@ -152,44 +170,76 @@ class TextStyleOps {
     return null;
   }
 
-  static FontWeight fontWeightTryParse(String value) {
-    switch (value) {
-      case kCssFontWeightBold:
-        return FontWeight.bold;
-      case kCssFontWeight100:
-        return FontWeight.w100;
-      case kCssFontWeight200:
-        return FontWeight.w200;
-      case kCssFontWeight300:
-        return FontWeight.w300;
-      case kCssFontWeight400:
-        return FontWeight.w400;
-      case kCssFontWeight500:
-        return FontWeight.w500;
-      case kCssFontWeight600:
-        return FontWeight.w600;
-      case kCssFontWeight700:
-        return FontWeight.w700;
-      case kCssFontWeight800:
-        return FontWeight.w800;
-      case kCssFontWeight900:
-        return FontWeight.w900;
+  static FontWeight? fontWeightTryParse(css.Expression expression) {
+    if (expression is css.LiteralTerm) {
+      if (expression is css.NumberTerm) {
+        switch (expression.number) {
+          case 100:
+            return FontWeight.w100;
+          case 200:
+            return FontWeight.w200;
+          case 300:
+            return FontWeight.w300;
+          case 400:
+            return FontWeight.w400;
+          case 500:
+            return FontWeight.w500;
+          case 600:
+            return FontWeight.w600;
+          case 700:
+            return FontWeight.w700;
+          case 800:
+            return FontWeight.w800;
+          case 900:
+            return FontWeight.w900;
+        }
+      }
+
+      switch (expression.valueAsString) {
+        case kCssFontWeightBold:
+          return FontWeight.bold;
+      }
     }
 
     return null;
   }
 
-  static double _fontSizeTryParse(WidgetFactory wf, TextStyleHtml p, String v) {
+  static TextStyleHtml whitespace(TextStyleHtml p, CssWhitespace v) =>
+      p.copyWith(whitespace: v);
+
+  static CssWhitespace? whitespaceTryParse(String value) {
+    switch (value) {
+      case kCssWhitespacePre:
+        return CssWhitespace.pre;
+      case kCssWhitespaceNormal:
+        return CssWhitespace.normal;
+    }
+
+    return null;
+  }
+
+  static double? _fontSizeTryParse(TextStyleHtml p, css.Expression v) {
     final length = tryParseCssLength(v);
     if (length != null) {
-      final lengthValue = length.getValue(
-        p,
-        baseValue: p.parent?.style?.fontSize,
-        scaleFactor: p.getDependency<MediaQueryData>().textScaleFactor,
-      );
+      final lengthValue = _fontSizeTryParseCssLength(p, length);
       if (lengthValue != null) return lengthValue;
     }
 
+    if (v is css.LiteralTerm) {
+      return _fontSizeTryParseTerm(p, v.valueAsString);
+    }
+
+    return null;
+  }
+
+  static double? _fontSizeTryParseCssLength(TextStyleHtml p, CssLength v) =>
+      v.getValue(
+        p,
+        baseValue: p.parent?.style.fontSize,
+        scaleFactor: p.getDependency<MediaQueryData>().textScaleFactor,
+      );
+
+  static double? _fontSizeTryParseTerm(TextStyleHtml p, String v) {
     switch (v) {
       case kCssFontSizeXxLarge:
         return _fontSizeMultiplyRootWith(p, 2.0);
@@ -207,55 +257,65 @@ class TextStyleOps {
         return _fontSizeMultiplyRootWith(p, .5625);
 
       case kCssFontSizeLarger:
-        return _fontSizeMultiplyWith(p.parent?.style?.fontSize, 1.2);
+        return _fontSizeMultiplyWith(p.parent?.style.fontSize, 1.2);
       case kCssFontSizeSmaller:
-        return _fontSizeMultiplyWith(p.parent?.style?.fontSize, 15 / 18);
+        return _fontSizeMultiplyWith(p.parent?.style.fontSize, 15 / 18);
     }
 
     return null;
   }
 
-  static double _fontSizeMultiplyRootWith(TextStyleHtml tsh, double value) {
+  static double? _fontSizeMultiplyRootWith(TextStyleHtml tsh, double value) {
     var root = tsh;
     while (root.parent != null) {
-      root = root.parent;
+      root = root.parent!;
     }
 
     return _fontSizeMultiplyWith(root.style.fontSize, value);
   }
 
-  static double _fontSizeMultiplyWith(double fontSize, double value) =>
+  static double? _fontSizeMultiplyWith(double? fontSize, double value) =>
       fontSize != null ? fontSize * value : null;
 
-  static double _lineHeightTryParse(
-      WidgetFactory wf, TextStyleHtml p, String v) {
-    if (v == kCssLineHeightNormal) return -1;
+  static double? _lineHeightTryParse(
+      WidgetFactory wf, TextStyleHtml p, css.Expression v) {
+    if (v is css.LiteralTerm) {
+      if (v is css.NumberTerm) {
+        final number = v.number.toDouble();
+        if (number > 0) return number;
+      }
 
-    final number = double.tryParse(v);
-    if (number != null && number > 0) return number;
+      switch (v.valueAsString) {
+        case kCssLineHeightNormal:
+          return -1;
+      }
+    }
+
+    final fontSize = p.style.fontSize;
+    if (fontSize == null) return null;
 
     final length = tryParseCssLength(v);
     if (length == null) return null;
 
     final lengthValue = length.getValue(
       p,
-      baseValue: p.style.fontSize,
+      baseValue: fontSize,
       scaleFactor: p.getDependency<MediaQueryData>().textScaleFactor,
     );
     if (lengthValue == null) return null;
 
-    return lengthValue / p.style.fontSize;
+    return lengthValue / fontSize;
   }
 }
 
 @immutable
 class TextDeco {
-  final Color color;
-  final bool over;
-  final bool strike;
-  final TextDecorationStyle style;
-  final CssLength thickness;
-  final bool under;
+  final Color? color;
+  final bool? over;
+  final bool? strike;
+  final TextDecorationStyle? style;
+  final CssLength? thickness;
+  final bool? under;
 
   TextDeco({
     this.color,
@@ -266,17 +326,14 @@ class TextDeco {
     this.under,
   });
 
-  factory TextDeco.tryParse(String value) {
-    for (final v in splitCssValues(value)) {
-      switch (v) {
+  static TextDeco? tryParse(List<css.Expression> expressions) {
+    for (final expression in expressions) {
+      if (expression is! css.LiteralTerm) continue;
+      switch (expression.valueAsString) {
         case kCssTextDecorationLineThrough:
           return TextDeco(strike: true);
         case kCssTextDecorationNone:
-          return TextDeco(
-            over: false,
-            strike: false,
-            under: false,
-          );
+          return TextDeco(over: false, strike: false, under: false);
         case kCssTextDecorationOverline:
           return TextDeco(over: true);
         case kCssTextDecorationUnderline:
