@@ -1,3 +1,4 @@
+import 'package:demo_app/model/show_perf_overlay.dart';
 import 'package:flutter/material.dart';
 
 import 'custom_styles_builder.dart';
@@ -6,9 +7,11 @@ import 'font_size.dart';
 import 'golden.dart';
 import 'hello_world.dart';
 import 'hello_world_core.dart';
+import 'huge_html.dart';
 import 'iframe.dart';
-import 'iframe_twitter.dart';
 import 'img.dart';
+import 'img_file.dart';
+import 'photo_view.dart';
 import 'smilie.dart';
 import 'video.dart';
 import 'wordpress.dart';
@@ -18,20 +21,27 @@ class HomeScreen extends StatelessWidget {
     'Hello World': () => HelloWorldScreen(),
     'Hello World (core)': () => HelloWorldCoreScreen(),
     'Iframe': () => IframeScreen(),
-    'Iframe/Twitter': () => IframeTwitterScreen(),
     'Images': () => ImgScreen(),
+    'Image (file://)': () => ImgFileScreen(),
     'Video': () => VideoScreen(),
     'customStylesBuilder': () => CustomStylesBuilderScreen(),
     'customWidgetBuilder': () => CustomWidgetBuilderScreen(),
     'font-size': () => FontSizeScreen(),
     'Goldens': () => GoldensScreen(),
+    'HugeHtml': () => HugeHtmlScreen(),
+    'Photo View': () => PhotoViewScreen(),
     'Smilie': () => SmilieScreen(),
     'Wordpress': () => WordpressScreen(),
   };
 
   @override
   Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: Text('Demo app')),
+        appBar: AppBar(
+          title: Text('Demo app'),
+          actions: [
+            ShowPerfIconButton(),
+          ],
+        ),
         body: ListView(
           children: _screens.keys
               .map((title) => ListTile(
