@@ -1,5 +1,4 @@
-import 'dart:math';
-
+import 'package:csslib/visitor.dart' as css;
 import 'package:flutter/widgets.dart';
 import 'package:html/dom.dart' as dom;
 
@@ -9,7 +8,6 @@ import 'core_widget_factory.dart';
 part 'data/build_bits.dart';
 part 'data/css.dart';
 part 'data/image.dart';
-part 'data/table.dart';
 part 'data/text_style.dart';
 
 /// A building element metadata.
@@ -17,68 +15,62 @@ abstract class BuildMetadata {
   /// The associatd element.
   final dom.Element element;
 
-  final TextStyleBuilder _tsb;
+  /// The associated [TextStyleBuilder].
+  final TextStyleBuilder tsb;
 
   /// Creates a node.
-  BuildMetadata(this.element, this._tsb);
+  BuildMetadata(this.element, this.tsb);
 
   /// The registered build ops.
   Iterable<BuildOp> get buildOps;
 
-  /// Returns `true` if node should be rendered as block.
-  bool get isBlockElement;
-
   /// The parents' build ops that have [BuildOp.onChild].
   Iterable<BuildOp> get parentOps;
 
-  /// The inline styles.
+  /// The styling declarations.
   ///
-  /// These are usually collected from:
+  /// These are collected from:
   ///
   /// - [WidgetFactory.parse] or [BuildOp.onChild] by calling `meta[key] = value`
   /// - [BuildOp.defaultStyles] returning a map
   /// - Attribute `style` of [domElement]
-  Iterable<MapEntry<String, String>> get styles;
+  List<css.Declaration> get styles;
 
-  /// Sets whether node should be rendered as block.
-  set isBlockElement(bool value);
+  /// Returns `true` if subtree will be built.
+  ///
+  /// May returns `null` if metadata is still being collected.
+  /// There are a few things that may trigger subtree building:
+  /// - Some [BuildOp] has a mandatory `onWidgets` callback
+  /// - Inline style `display: block`
+  ///
+  /// See [BuildOp.onWidgetsIsOptional].
+  bool? get willBuildSubtree;
 
   /// Adds an inline style.
   operator []=(String key, String value);
 
-  /// Gets an inline style value by key.
-  String operator [](String key) {
-    String value;
-    for (final x in styles) {
-      if (x.key == key) value = x.value;
+  /// Gets a styling declaration by `property`.
+  css.Declaration? operator [](String key) {
+    for (final style in styles.reversed) {
+      if (style.property == key) return style;
     }
-    return value;
+    return null;
   }
 
   /// Registers a build op.
   void register(BuildOp op);
 
   @override
-  String toString() =>
-      'BuildMetadata(${element == null ? "root" : element.outerHtml})';
-
-  /// Enqueues a text style builder callback.
-  ///
-  /// Returns the associated [TextStyleBuilder].
-  TextStyleBuilder tsb<T>([
-    TextStyleHtml Function(TextStyleHtml tsh, T input) builder,
-    T input,
-  ]) =>
-      _tsb..enqueue(builder, input);
+  String toString() => 'BuildMetadata(${element.outerHtml})';
 }
 
 /// A building operation to customize how a DOM element is rendered.
 @immutable
 class BuildOp {
-  /// Controls whether the element should be rendered with [CssBlock].
+  /// The recommended maximum value for [priority].
   ///
-  /// Default: `true` if [onWidgets] callback is set, `false` otherwise.
-  final bool isBlockElement;
+  /// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER
+  static const kPriorityMax = 9007199254740991;
 
   /// The execution priority, op with lower priority will run first.
   ///
@@ -98,7 +90,7 @@ class BuildOp {
   ///
   /// Note: op must be registered early for this to work e.g.
   /// in [WidgetFactory.parse] or [onChild].
-  final Map<String, String> Function(dom.Element element) defaultStyles;
+  final Map<String, String> Function(dom.Element element)? defaultStyles;
 
   /// The callback that will be called whenver a child element is found.
   ///
@@ -114,24 +106,29 @@ class BuildOp {
   /// );
   ///
   /// ```
-  final void Function(BuildMetadata childMeta) onChild;
+  final void Function(BuildMetadata childMeta)? onChild;
 
   /// The callback that will be called when child elements have been processed.
-  final void Function(BuildMetadata meta, BuildTree tree) onTree;
+  final void Function(BuildMetadata meta, BuildTree tree)? onTree;
 
   /// The callback that will be called when child elements have been built.
   ///
   /// Note: only works if it's a block element.
-  final Iterable<Widget> Function(
-      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets) onWidgets;
+  final Iterable<Widget>? Function(
+      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets)? onWidgets;
+
+  /// Controls whether the element should be forced to be rendered as block.
+  ///
+  /// Default: `false`.
+  final bool onWidgetsIsOptional;
 
   /// Creates a build op.
   BuildOp({
     this.defaultStyles,
-    bool isBlockElement,
     this.onChild,
     this.onTree,
     this.onWidgets,
+    this.onWidgetsIsOptional = false,
     this.priority = 10,
-  }) : isBlockElement = isBlockElement ?? onWidgets != null;
+  });
 }
