@@ -1,18 +1,21 @@
 import 'dart:math';
 
+import 'package:csslib/parser.dart' as css;
+import 'package:csslib/visitor.dart' as css;
 import 'package:flutter/widgets.dart';
 
 import '../core_data.dart';
+import '../core_helpers.dart';
 
 part 'parser/border.dart';
 part 'parser/color.dart';
 part 'parser/length.dart';
 
-final _spacingRegExp = RegExp(r'\s+');
-Iterable<String> splitCssValues(String value) => value.split(_spacingRegExp);
-
-final _attrStyleRegExp = RegExp(r'([a-zA-Z\-]+)\s*:\s*([^;]*)');
-Iterable<MapEntry<String, String>> splitAttributeStyle(String value) =>
-    _attrStyleRegExp
-        .allMatches(value)
-        .map((m) => MapEntry(m[1].trim(), m[2].trim()));
+const kSuffixBlockEnd = '-block-end';
+const kSuffixBlockStart = '-block-start';
+const kSuffixBottom = '-bottom';
+const kSuffixInlineEnd = '-inline-end';
+const kSuffixInlineStart = '-inline-start';
+const kSuffixLeft = '-left';
+const kSuffixRight = '-right';
+const kSuffixTop = '-top';
