@@ -10,20 +10,21 @@ WidgetPlaceholder _paddingInlineBefore(TextStyleBuilder tsb, CssLengthBox b) =>
     WidgetPlaceholder<CssLengthBox>(b).wrapWith((context, _) =>
         _paddingInlineSizedBox(b.getValueLeft(tsb.build(context))));
 
-Widget _paddingInlineSizedBox(double width) =>
+Widget _paddingInlineSizedBox(double? width) =>
     width != null && width > 0 ? SizedBox(width: width) : widget0;
 
 class StylePadding {
+  static const kPriorityBoxModel3k = 3000;
+
   final WidgetFactory wf;
 
   StylePadding(this.wf);
 
   BuildOp get buildOp => BuildOp(
-        isBlockElement: false,
         onTree: (meta, tree) {
-          if (meta.isBlockElement) return;
+          if (meta.willBuildSubtree == true) return;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
-          if (padding?.hasLeftOrRight != true) return;
+          if (padding == null || !padding.hasPositiveLeftOrRight) return;
 
           return wrapTree(
             tree,
@@ -34,28 +35,36 @@ class StylePadding {
           );
         },
         onWidgets: (meta, widgets) {
-          if (widgets?.isNotEmpty != true) return null;
+          if (meta.willBuildSubtree == false) return widgets;
+          if (widgets.isEmpty) return widgets;
+
           final padding = tryParseCssLengthBox(meta, kCssPadding);
           if (padding == null) return null;
 
-          return listOrNull(wf
-              .buildColumnPlaceholder(meta, widgets)
-              ?.wrapWith((c, w) => _build(c, meta, w, padding)));
+          return [
+            WidgetPlaceholder(
+              padding,
+              child: wf.buildColumnPlaceholder(meta, widgets),
+            ).wrapWith(
+              (context, child) => _build(meta, context, child, padding),
+            )
+          ];
         },
-        priority: 9999,
+        onWidgetsIsOptional: true,
+        priority: kPriorityBoxModel3k,
       );
 
-  Widget _build(BuildContext context, BuildMetadata meta, Widget child,
+  Widget? _build(BuildMetadata meta, BuildContext context, Widget child,
       CssLengthBox padding) {
-    final tsh = meta.tsb().build(context);
+    final tsh = meta.tsb.build(context);
     return wf.buildPadding(
       meta,
       child,
       EdgeInsets.fromLTRB(
-        padding.getValueLeft(tsh) ?? 0,
-        padding.top?.getValue(tsh) ?? 0,
-        padding.getValueRight(tsh) ?? 0,
-        padding.bottom?.getValue(tsh) ?? 0,
+        max(padding.getValueLeft(tsh) ?? 0.0, 0.0),
+        max(padding.top?.getValue(tsh) ?? 0.0, 0.0),
+        max(padding.getValueRight(tsh) ?? 0.0, 0.0),
+        max(padding.bottom?.getValue(tsh) ?? 0.0, 0.0),
       ),
     );
   }
