@@ -1,26 +1,22 @@
 part of '../core_ops.dart';
 
 const kAttributeAHref = 'href';
+const kAttributeAName = 'name';
 const kTagA = 'a';
 
 class TagA {
   final WidgetFactory wf;
-  final Color Function() colorCallback;
 
-  TagA(this.wf, this.colorCallback);
+  TagA(this.wf);
 
   BuildOp get buildOp => BuildOp(
         defaultStyles: (_) {
           final styles = {kCssTextDecoration: kCssTextDecorationUnderline};
 
-          final color = colorCallback?.call();
-          if (color != null) styles[kCssColor] = convertColorToHex(color);
-
           return styles;
         },
-        isBlockElement: false,
         onTree: (meta, tree) {
-          if (meta.isBlockElement) return;
+          if (meta.willBuildSubtree == true) return;
 
           final onTap = _gestureTapCallback(meta);
           if (onTap == null) return;
@@ -29,37 +25,42 @@ class TagA {
             if (bit is WidgetBit) {
               bit.child.wrapWith(
                   (_, child) => wf.buildGestureDetector(meta, child, onTap));
-            } else if (bit.tsb != null) {
+            } else if (bit is! WhitespaceBit) {
               _TagABit(bit.parent, bit.tsb, onTap).insertAfter(bit);
             }
           }
         },
         onWidgets: (meta, widgets) {
+          if (meta.willBuildSubtree == false) return widgets;
+
           final onTap = _gestureTapCallback(meta);
           if (onTap == null) return widgets;
 
           return listOrNull(wf.buildColumnPlaceholder(meta, widgets)?.wrapWith(
               (_, child) => wf.buildGestureDetector(meta, child, onTap)));
         },
+        onWidgetsIsOptional: true,
       );
 
-  GestureTapCallback _gestureTapCallback(BuildMetadata meta) {
+  GestureTapCallback? _gestureTapCallback(BuildMetadata meta) {
     final href = meta.element.attributes[kAttributeAHref];
-    return wf.gestureTapCallback(wf.urlFull(href) ?? href);
+    return href != null
+        ? wf.gestureTapCallback(wf.urlFull(href) ?? href)
+        : null;
   }
 }
 
-class _TagABit extends BuildBit<GestureRecognizer, GestureRecognizer> {
+class _TagABit extends BuildBit<GestureRecognizer?, GestureRecognizer> {
   final GestureTapCallback onTap;
 
-  _TagABit(BuildTree parent, TextStyleBuilder tsb, this.onTap)
+  _TagABit(BuildTree? parent, TextStyleBuilder tsb, this.onTap)
       : super(parent, tsb);
 
   @override
-  bool get swallowWhitespace => null;
+  bool? get swallowWhitespace => null;
 
   @override
-  GestureRecognizer buildBit(GestureRecognizer recognizer) {
+  GestureRecognizer buildBit(GestureRecognizer? recognizer) {
     if (recognizer is TapGestureRecognizer) {
       recognizer.onTap = onTap;
       return recognizer;
@@ -69,6 +70,6 @@ class _TagABit extends BuildBit<GestureRecognizer, GestureRecognizer> {
   }
 
   @override
-  BuildBit copyWith({BuildTree parent, TextStyleBuilder tsb}) =>
+  BuildBit copyWith({BuildTree? parent, TextStyleBuilder? tsb}) =>
       _TagABit(parent ?? this.parent, tsb ?? this.tsb, onTap);
 }
