@@ -2,187 +2,52 @@ part of '../core_parser.dart';
 
 const kCssColor = 'color';
 
-// https://drafts.csswg.org/css-color/#named-colors
-/*
-(function() {
-    var colors = '';
-    var count = 0;
-    document.querySelectorAll('.named-color-table tr').forEach((tr) => {
-        var dfn = tr.querySelector('th > dfn');
-        var td = tr.querySelector('td:nth-child(4)');
-        if (!dfn || !td) return;
-        var color = dfn.textContent.trim();
-        var hex = td.textContent.trim();
-        colors += `  '${color}': '${hex}',\n`;
-        count++;
-    });
-    console.log(`const kCssColors = {\n${colors}};`);
-    return count;
-})();
-*/
-const kCssColors = {
-  'aliceblue': '#f0f8ff',
-  'antiquewhite': '#faebd7',
-  'aqua': '#00ffff',
-  'aquamarine': '#7fffd4',
-  'azure': '#f0ffff',
-  'beige': '#f5f5dc',
-  'bisque': '#ffe4c4',
-  'black': '#000000',
-  'blanchedalmond': '#ffebcd',
-  'blue': '#0000ff',
-  'blueviolet': '#8a2be2',
-  'brown': '#a52a2a',
-  'burlywood': '#deb887',
-  'cadetblue': '#5f9ea0',
-  'chartreuse': '#7fff00',
-  'chocolate': '#d2691e',
-  'coral': '#ff7f50',
-  'cornflowerblue': '#6495ed',
-  'cornsilk': '#fff8dc',
-  'crimson': '#dc143c',
-  'cyan': '#00ffff',
-  'darkblue': '#00008b',
-  'darkcyan': '#008b8b',
-  'darkgoldenrod': '#b8860b',
-  'darkgray': '#a9a9a9',
-  'darkgreen': '#006400',
-  'darkgrey': '#a9a9a9',
-  'darkkhaki': '#bdb76b',
-  'darkmagenta': '#8b008b',
-  'darkolivegreen': '#556b2f',
-  'darkorange': '#ff8c00',
-  'darkorchid': '#9932cc',
-  'darkred': '#8b0000',
-  'darksalmon': '#e9967a',
-  'darkseagreen': '#8fbc8f',
-  'darkslateblue': '#483d8b',
-  'darkslategray': '#2f4f4f',
-  'darkslategrey': '#2f4f4f',
-  'darkturquoise': '#00ced1',
-  'darkviolet': '#9400d3',
-  'deeppink': '#ff1493',
-  'deepskyblue': '#00bfff',
-  'dimgray': '#696969',
-  'dimgrey': '#696969',
-  'dodgerblue': '#1e90ff',
-  'firebrick': '#b22222',
-  'floralwhite': '#fffaf0',
-  'forestgreen': '#228b22',
-  'fuchsia': '#ff00ff',
-  'gainsboro': '#dcdcdc',
-  'ghostwhite': '#f8f8ff',
-  'gold': '#ffd700',
-  'goldenrod': '#daa520',
-  'gray': '#808080',
-  'green': '#008000',
-  'greenyellow': '#adff2f',
-  'grey': '#808080',
-  'honeydew': '#f0fff0',
-  'hotpink': '#ff69b4',
-  'indianred': '#cd5c5c',
-  'indigo': '#4b0082',
-  'ivory': '#fffff0',
-  'khaki': '#f0e68c',
-  'lavender': '#e6e6fa',
-  'lavenderblush': '#fff0f5',
-  'lawngreen': '#7cfc00',
-  'lemonchiffon': '#fffacd',
-  'lightblue': '#add8e6',
-  'lightcoral': '#f08080',
-  'lightcyan': '#e0ffff',
-  'lightgoldenrodyellow': '#fafad2',
-  'lightgray': '#d3d3d3',
-  'lightgreen': '#90ee90',
-  'lightgrey': '#d3d3d3',
-  'lightpink': '#ffb6c1',
-  'lightsalmon': '#ffa07a',
-  'lightseagreen': '#20b2aa',
-  'lightskyblue': '#87cefa',
-  'lightslategray': '#778899',
-  'lightslategrey': '#778899',
-  'lightsteelblue': '#b0c4de',
-  'lightyellow': '#ffffe0',
-  'lime': '#00ff00',
-  'limegreen': '#32cd32',
-  'linen': '#faf0e6',
-  'magenta': '#ff00ff',
-  'maroon': '#800000',
-  'mediumaquamarine': '#66cdaa',
-  'mediumblue': '#0000cd',
-  'mediumorchid': '#ba55d3',
-  'mediumpurple': '#9370db',
-  'mediumseagreen': '#3cb371',
-  'mediumslateblue': '#7b68ee',
-  'mediumspringgreen': '#00fa9a',
-  'mediumturquoise': '#48d1cc',
-  'mediumvioletred': '#c71585',
-  'midnightblue': '#191970',
-  'mintcream': '#f5fffa',
-  'mistyrose': '#ffe4e1',
-  'moccasin': '#ffe4b5',
-  'navajowhite': '#ffdead',
-  'navy': '#000080',
-  'oldlace': '#fdf5e6',
-  'olive': '#808000',
-  'olivedrab': '#6b8e23',
-  'orange': '#ffa500',
-  'orangered': '#ff4500',
-  'orchid': '#da70d6',
-  'palegoldenrod': '#eee8aa',
-  'palegreen': '#98fb98',
-  'paleturquoise': '#afeeee',
-  'palevioletred': '#db7093',
-  'papayawhip': '#ffefd5',
-  'peachpuff': '#ffdab9',
-  'peru': '#cd853f',
-  'pink': '#ffc0cb',
-  'plum': '#dda0dd',
-  'powderblue': '#b0e0e6',
-  'purple': '#800080',
-  'rebeccapurple': '#663399',
-  'red': '#ff0000',
-  'rosybrown': '#bc8f8f',
-  'royalblue': '#4169e1',
-  'saddlebrown': '#8b4513',
-  'salmon': '#fa8072',
-  'sandybrown': '#f4a460',
-  'seagreen': '#2e8b57',
-  'seashell': '#fff5ee',
-  'sienna': '#a0522d',
-  'silver': '#c0c0c0',
-  'skyblue': '#87ceeb',
-  'slateblue': '#6a5acd',
-  'slategray': '#708090',
-  'slategrey': '#708090',
-  'snow': '#fffafa',
-  'springgreen': '#00ff7f',
-  'steelblue': '#4682b4',
-  'tan': '#d2b48c',
-  'teal': '#008080',
-  'thistle': '#d8bfd8',
-  'tomato': '#ff6347',
-  'turquoise': '#40e0d0',
-  'violet': '#ee82ee',
-  'wheat': '#f5deb3',
-  'white': '#ffffff',
-  'whitesmoke': '#f5f5f5',
-  'yellow': '#ffff00',
-  'yellowgreen': '#9acd32',
-};
+Color? tryParseColor(css.Expression? expression) {
+  if (expression == null) return null;
 
-final _colorHexRegExp = RegExp(r'^#([a-f0-9]{3,8})$', caseSensitive: false);
-final _colorHslRegExp = RegExp(
-    r'^(hsla?)\(([0-9.-]+)(deg|rad|grad|turn)?[,\s]+([0-9.]+%)[,\s]+([0-9.]+%)([,\s/]+([0-9.]+%?))?\)$');
-final _colorRgbRegExp = RegExp(
-    r'^(rgba?)\(([0-9.]+%?)[,\s]+([0-9.]+%?)[,\s]+([0-9.]+%?)([,\s/]+([0-9.]+%?))?\)$');
-
-Color tryParseColor(String value) {
-  if (value == null) return null;
-
-  final hexMatch = _colorHexRegExp.firstMatch(value);
-  if (hexMatch != null) {
-    final hex = hexMatch[1].toUpperCase();
+  if (expression is css.FunctionTerm) {
+    switch (expression.text) {
+      case 'hsl':
+      case 'hsla':
+        final params = expression.params;
+        if (params.length >= 3) {
+          final param0 = params[0];
+          final h = param0 is css.NumberTerm
+              ? _parseColorHue(param0.number)
+              : param0 is css.AngleTerm
+                  ? _parseColorHue(param0.value, param0.unit)
+                  : null;
+          final param1 = params[1];
+          final s = param1 is css.PercentageTerm
+              ? param1.valueAsDouble.clamp(0.0, 1.0)
+              : null;
+          final param2 = params[2];
+          final l = param2 is css.PercentageTerm
+              ? param2.valueAsDouble.clamp(0.0, 1.0)
+              : null;
+          final hslA = params.length >= 4 ? _parseColorAlpha(params[3]) : 1.0;
+          if (h != null && s != null && l != null && hslA != null) {
+            return HSLColor.fromAHSL(hslA, h, s, l).toColor();
+          }
+        }
+        break;
+      case 'rgb':
+      case 'rgba':
+        final params = expression.params;
+        if (params.length >= 3) {
+          final r = _parseColorRgbElement(params[0]);
+          final g = _parseColorRgbElement(params[1]);
+          final b = _parseColorRgbElement(params[2]);
+          final rgbA = params.length >= 4 ? _parseColorAlpha(params[3]) : 1.0;
+          if (r != null && g != null && b != null && rgbA != null) {
+            return Color.fromARGB((rgbA * 255).ceil(), r, g, b);
+          }
+        }
+        break;
+    }
+  } else if (expression is css.HexColorTerm) {
+    // cannot use expression.value directory due to issue with #f00 etc.
+    final hex = expression.text.toUpperCase();
     switch (hex.length) {
       case 3:
         return Color(int.parse('0xFF${_x2(hex)}'));
@@ -197,60 +62,38 @@ Color tryParseColor(String value) {
         final rgb = hex.substring(0, 6);
         return Color(int.parse('0x$alpha$rgb'));
     }
-  }
-
-  final valueLowerCase = value.toLowerCase();
-  if (valueLowerCase == 'transparent') return Color(0x00000000);
-  // TODO: add support for `currentcolor`
-  if (kCssColors.containsKey(valueLowerCase)) {
-    return tryParseColor(kCssColors[valueLowerCase]);
-  }
-
-  final rgbMatch = _colorRgbRegExp.firstMatch(valueLowerCase);
-  if (rgbMatch != null) {
-    final rgbR = _parseColorPart(rgbMatch[2], 0, 255);
-    final rgbG = _parseColorPart(rgbMatch[3], 0, 255);
-    final rgbB = _parseColorPart(rgbMatch[4], 0, 255);
-    final rgbA = _parseColorPart(rgbMatch[6] ?? '1', 0, 1);
-    if (rgbR != null && rgbG != null && rgbB != null && rgbA != null) {
-      return Color.fromARGB(
-        (255 * rgbA).ceil(),
-        rgbR.toInt(),
-        rgbG.toInt(),
-        rgbB.toInt(),
-      );
-    }
-  }
-
-  final hslMatch = _colorHslRegExp.firstMatch(valueLowerCase);
-  if (hslMatch != null) {
-    final hslH = _parseColorHue(hslMatch[2], hslMatch[3]);
-    final hslS = _parseColorPart(hslMatch[4], 0, 1);
-    final hslL = _parseColorPart(hslMatch[5], 0, 1);
-    final hslA = _parseColorPart(hslMatch[7] ?? '1', 0, 1);
-    if (hslH != null && hslS != null && hslL != null && hslA != null) {
-      return HSLColor.fromAHSL(hslA, hslH, hslS, hslL).toColor();
+  } else if (expression is css.LiteralTerm) {
+    switch (expression.valueAsString) {
+      // TODO: add support for `currentcolor`
+      case 'transparent':
+        return Color(0x00000000);
     }
   }
 
   return null;
 }
 
-double _parseColorHue(String number, String unit) {
-  final v = double.tryParse(number);
-  if (v == null) return null;
+double? _parseColorAlpha(css.Expression v) => (v is css.NumberTerm
+        ? v.number.toDouble()
+        : v is css.PercentageTerm
+            ? v.valueAsDouble
+            : null)
+    ?.clamp(0.0, 1.0);
+
+double _parseColorHue(num number, [int? unit]) {
+  final v = number is double ? number : number.toDouble();
 
   double deg;
   switch (unit) {
-    case 'rad':
+    case css.TokenKind.UNIT_ANGLE_RAD:
       final rad = v;
       deg = rad * (180 / pi);
       break;
-    case 'grad':
+    case css.TokenKind.UNIT_ANGLE_GRAD:
       final grad = v;
       deg = grad * 0.9;
       break;
-    case 'turn':
+    case css.TokenKind.UNIT_ANGLE_TURN:
       final turn = v;
       deg = turn * 360;
       break;
@@ -265,20 +108,12 @@ double _parseColorHue(String number, String unit) {
   return deg % 360;
 }
 
-double _parseColorPart(String value, double min, double max) {
-  double v;
-
-  if (value.endsWith('%')) {
-    final p = double.tryParse(value.substring(0, value.length - 1));
-    if (p == null || p < 0) return null;
-    v = p / 100.0 * max;
-  }
-
-  v ??= double.tryParse(value);
-
-  if (v < min || v > max) return null;
-  return v;
-}
+int? _parseColorRgbElement(css.Expression v) => (v is css.NumberTerm
+        ? v.number.ceil()
+        : v is css.PercentageTerm
+            ? (v.valueAsDouble * 255.0).ceil()
+            : null)
+    ?.clamp(0, 255);
 
 String _x2(String value) {
   final sb = StringBuffer();
