@@ -1,124 +1,111 @@
 part of '../core_parser.dart';
 
-const kCssLengthBoxSuffixBlockEnd = '-block-end';
-const kCssLengthBoxSuffixBlockStart = '-block-start';
-const kCssLengthBoxSuffixBottom = '-bottom';
-const kCssLengthBoxSuffixInlineEnd = '-inline-end';
-const kCssLengthBoxSuffixInlineStart = '-inline-start';
-const kCssLengthBoxSuffixLeft = '-left';
-const kCssLengthBoxSuffixRight = '-right';
-const kCssLengthBoxSuffixTop = '-top';
+CssLength? tryParseCssLength(css.Expression expression) {
+  if (expression is css.EmTerm) {
+    return CssLength((expression.value as num).toDouble(), CssLengthUnit.em);
+  } else if (expression is css.LengthTerm) {
+    final number = (expression.value as num).toDouble();
+    switch (expression.unit) {
+      case css.TokenKind.UNIT_LENGTH_PT:
+        return CssLength(number, CssLengthUnit.pt);
+      case css.TokenKind.UNIT_LENGTH_PX:
+        return CssLength(number, CssLengthUnit.px);
+    }
+  } else if (expression is css.LiteralTerm) {
+    if (expression is css.NumberTerm) {
+      if (expression.number == 0) return CssLength(0);
+    } else if (expression is css.PercentageTerm) {
+      return CssLength(
+          (expression.value as num).toDouble(), CssLengthUnit.percentage);
+    }
 
-final _cssLengthValues4 =
-    RegExp(r'^([^\s]+)\s+([^\s]+)\s+([^\s]+)\s+([^\s]+)$');
-final _cssLengthValues2 = RegExp(r'^([^\s]+)\s+([^\s]+)$');
-final _cssLengthValue = RegExp(r'^([\d\.]+)(em|%|pt|px)$');
-
-CssLength tryParseCssLength(String value) {
-  if (value == null) return null;
-  if (value == '0') return CssLength(0);
-
-  final match = _cssLengthValue.firstMatch(value);
-  if (match == null) return null;
-
-  final number = double.tryParse(match[1]);
-  if (number == null) return null;
-
-  switch (match[2]) {
-    case 'em':
-      return CssLength(number, CssLengthUnit.em);
-    case '%':
-      return CssLength(number, CssLengthUnit.percentage);
-    case 'pt':
-      return CssLength(number, CssLengthUnit.pt);
-    case 'px':
-      return CssLength(number, CssLengthUnit.px);
+    switch (expression.valueAsString) {
+      case 'auto':
+        return CssLength(1, CssLengthUnit.auto);
+    }
   }
 
   return null;
 }
 
-CssLengthBox tryParseCssLengthBox(BuildMetadata meta, String key) {
-  CssLengthBox output;
-
-  for (final style in meta.styles) {
-    if (!style.key.startsWith(key)) continue;
-
-    final suffix = style.key.substring(key.length);
-    switch (suffix) {
-      case '':
-        output = _parseCssLengthBoxAll(style.value);
-        break;
-
-      case kCssLengthBoxSuffixBlockEnd:
-      case kCssLengthBoxSuffixBlockStart:
-      case kCssLengthBoxSuffixBottom:
-      case kCssLengthBoxSuffixInlineEnd:
-      case kCssLengthBoxSuffixInlineStart:
-      case kCssLengthBoxSuffixLeft:
-      case kCssLengthBoxSuffixRight:
-      case kCssLengthBoxSuffixTop:
-        output = _parseCssLengthBoxOne(output, suffix, style.value);
-        break;
-    }
+CssLengthBox? _parseCssLengthBoxAll(List<css.Expression> expressions) {
+  switch (expressions.length) {
+    case 4:
+      return CssLengthBox(
+        top: tryParseCssLength(expressions[0]),
+        inlineEnd: tryParseCssLength(expressions[1]),
+        bottom: tryParseCssLength(expressions[2]),
+        inlineStart: tryParseCssLength(expressions[3]),
+      );
+    case 2:
+      final topBottom = tryParseCssLength(expressions[0]);
+      final leftRight = tryParseCssLength(expressions[1]);
+      return CssLengthBox(
+        top: topBottom,
+        inlineEnd: leftRight,
+        bottom: topBottom,
+        inlineStart: leftRight,
+      );
+    case 1:
+      final all = tryParseCssLength(expressions[0]);
+      return CssLengthBox(
+        top: all,
+        inlineEnd: all,
+        bottom: all,
+        inlineStart: all,
+      );
   }
 
-  return output;
+  return null;
 }
 
-CssLengthBox _parseCssLengthBoxAll(String value) {
-  final valuesFour = _cssLengthValues4.firstMatch(value);
-  if (valuesFour != null) {
-    return CssLengthBox(
-      top: tryParseCssLength(valuesFour[1]),
-      inlineEnd: tryParseCssLength(valuesFour[2]),
-      bottom: tryParseCssLength(valuesFour[3]),
-      inlineStart: tryParseCssLength(valuesFour[4]),
-    );
-  }
-
-  final valuesTwo = _cssLengthValues2.firstMatch(value);
-  if (valuesTwo != null) {
-    final topBottom = tryParseCssLength(valuesTwo[1]);
-    final leftRight = tryParseCssLength(valuesTwo[2]);
-    return CssLengthBox(
-      top: topBottom,
-      inlineEnd: leftRight,
-      bottom: topBottom,
-      inlineStart: leftRight,
-    );
-  }
-
-  final all = tryParseCssLength(value);
-  return CssLengthBox(top: all, inlineEnd: all, bottom: all, inlineStart: all);
-}
-
-CssLengthBox _parseCssLengthBoxOne(
-  CssLengthBox existing,
+CssLengthBox? _parseCssLengthBoxOne(
+  CssLengthBox? existing,
   String suffix,
-  String value,
+  css.Expression expression,
 ) {
-  final parsed = tryParseCssLength(value);
+  final parsed = tryParseCssLength(expression);
   if (parsed == null) return existing;
 
   existing ??= CssLengthBox();
 
   switch (suffix) {
-    case kCssLengthBoxSuffixBottom:
-    case kCssLengthBoxSuffixBlockEnd:
+    case kSuffixBottom:
+    case kSuffixBlockEnd:
       return existing.copyWith(bottom: parsed);
-    case kCssLengthBoxSuffixInlineEnd:
+    case kSuffixInlineEnd:
       return existing.copyWith(inlineEnd: parsed);
-    case kCssLengthBoxSuffixInlineStart:
+    case kSuffixInlineStart:
       return existing.copyWith(inlineStart: parsed);
-    case kCssLengthBoxSuffixLeft:
+    case kSuffixLeft:
       return existing.copyWith(left: parsed);
-    case kCssLengthBoxSuffixRight:
+    case kSuffixRight:
       return existing.copyWith(right: parsed);
-    case kCssLengthBoxSuffixTop:
-    case kCssLengthBoxSuffixBlockStart:
+    case kSuffixTop:
+    case kSuffixBlockStart:
       return existing.copyWith(top: parsed);
   }
 
   return existing;
+}
+
+CssLengthBox? tryParseCssLengthBox(BuildMetadata meta, String prefix) {
+  CssLengthBox? output;
+
+  for (final style in meta.styles) {
+    final key = style.property;
+    if (!key.startsWith(prefix)) continue;
+
+    final suffix = key.substring(prefix.length);
+    if (suffix.isEmpty) {
+      output = _parseCssLengthBoxAll(style.values);
+    } else {
+      final expression = style.value;
+      if (expression != null) {
+        output = _parseCssLengthBoxOne(output, suffix, expression);
+      }
+    }
+  }
+
+  return output;
 }
