@@ -1,8 +1,10 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 import 'package:flutter_widget_from_html_core/src/internal/tsh_widget.dart';
-import 'package:network_image_mock/network_image_mock.dart';
+
+import 'package:mocktail_image_network/mocktail_image_network.dart';
 
 import '_.dart' as helper;
 
@@ -11,7 +13,7 @@ Future<String> explain(WidgetTester t, HtmlWidget hw) =>
 
 void main() {
   group('buildAsync', () {
-    final explain = (WidgetTester tester, String html, bool buildAsync) =>
+    final explain = (WidgetTester tester, String html, bool? buildAsync) =>
         tester.runAsync(() => helper.explain(tester, null,
             hw: HtmlWidget(
               html,
@@ -42,8 +44,8 @@ void main() {
     final explain = (
       WidgetTester tester,
       String html, {
-      AsyncWidgetBuilder<Widget> buildAsyncBuilder,
-      bool withData,
+      AsyncWidgetBuilder<Widget>? buildAsyncBuilder,
+      bool withData = true,
     }) =>
         tester.runAsync(() => helper.explain(tester, null,
             buildFutureBuilderWithData: withData,
@@ -61,7 +63,8 @@ void main() {
         expect(explained, equals('[FutureBuilder:[RichText:(:$html)]]'));
       });
 
-      testWidgets('renders indicator', (WidgetTester tester) async {
+      testWidgets('renders CircularProgressIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.android;
         final html = 'Foo';
         final explained = await explain(tester, html, withData: false);
         expect(
@@ -69,15 +72,30 @@ void main() {
             equals('[FutureBuilder:'
                 '[Center:child='
                 '[Padding:(8,8,8,8),child='
-                '[Text:Loading...]'
+                '[CircularProgressIndicator]'
                 ']]]'));
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('renders CupertinoActivityIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+        final html = 'Foo';
+        final explained = await explain(tester, html, withData: false);
+        expect(
+            explained,
+            equals('[FutureBuilder:'
+                '[Center:child='
+                '[Padding:(8,8,8,8),child='
+                '[CupertinoActivityIndicator]'
+                ']]]'));
+        debugDefaultTargetPlatformOverride = null;
       });
     });
 
     group('custom', () {
       final buildAsyncBuilder =
           (BuildContext _, AsyncSnapshot<Widget> snapshot) =>
-              snapshot.hasData ? snapshot.data : Text('No data');
+              snapshot.data ?? Text('No data');
 
       testWidgets('renders data', (WidgetTester tester) async {
         final html = 'Foo';
@@ -108,11 +126,11 @@ void main() {
       WidgetTester tester,
       String html,
       bool enableCaching, {
-      Uri baseUrl,
-      bool buildAsync,
-      Color hyperlinkColor = const Color.fromRGBO(0, 0, 255, 1),
-      RebuildTriggers rebuildTriggers,
-      TextStyle textStyle,
+      Uri? baseUrl,
+      bool? buildAsync,
+      Color? hyperlinkColor,
+      RebuildTriggers? rebuildTriggers,
+      TextStyle? textStyle,
     }) =>
         helper.explain(tester, null,
             hw: HtmlWidget(
@@ -126,7 +144,7 @@ void main() {
               textStyle: textStyle,
             ));
 
-    final _expect = (Widget built1, Widget built2, Matcher matcher) {
+    final _expect = (Widget? built1, Widget? built2, Matcher matcher) {
       final widget1 = (built1 as TshWidget).child;
       final widget2 = (built2 as TshWidget).child;
       expect(widget1 == widget2, matcher);
@@ -250,7 +268,7 @@ void main() {
 
     testWidgets(
       'renders with value',
-      (tester) => mockNetworkImagesFor(() async {
+      (tester) => mockNetworkImages(() async {
         final explained = await explain(
           tester,
           HtmlWidget(html, baseUrl: baseUrl, key: helper.hwKey),
@@ -258,9 +276,10 @@ void main() {
         expect(
             explained,
             equals(
+                '[CssSizing:height≥0.0,height=auto,width≥0.0,width=auto,child='
                 '[Image:image=NetworkImage("http://base.com/path/image.png", scale: 1.0),'
                 'semanticLabel=image dot png'
-                ']'));
+                ']]'));
       }),
     );
   });
@@ -289,15 +308,15 @@ void main() {
 
   group('customWidgetBuilder', () {
     final CustomWidgetBuilder customWidgetBuilder = (_) => Text('Bar');
-    final html = '<span>Foo</span>';
+    final html = 'Foo <span>bar</span>';
 
     testWidgets('renders without value', (WidgetTester tester) async {
       final e = await explain(tester, HtmlWidget(html, key: helper.hwKey));
-      expect(e, equals('[RichText:(:Foo)]'));
+      expect(e, equals('[RichText:(:Foo bar)]'));
     });
 
     testWidgets('renders with value', (WidgetTester tester) async {
-      final explained = await explain(
+      final e = await explain(
         tester,
         HtmlWidget(
           html,
@@ -305,7 +324,31 @@ void main() {
           key: helper.hwKey,
         ),
       );
-      expect(explained, equals('[CssBlock:child=[Text:Bar]]'));
+      expect(e, equals('[Column:children=[RichText:(:Foo)],[Text:Bar]]'));
+    });
+  });
+
+  group('customWidgetBuilder (TABLE)', () {
+    final CustomWidgetBuilder customWidgetBuilder =
+        (e) => e.localName == 'table' ? Text('Bar') : null;
+    final html = 'Foo <table><tr><td>bar</td></tr></table>';
+
+    testWidgets('renders without value', (WidgetTester tester) async {
+      final explained =
+          await explain(tester, HtmlWidget(html, key: helper.hwKey));
+      expect(explained, isNot(contains('Bar')));
+    });
+
+    testWidgets('renders with value', (WidgetTester tester) async {
+      final e = await explain(
+        tester,
+        HtmlWidget(
+          html,
+          customWidgetBuilder: customWidgetBuilder,
+          key: helper.hwKey,
+        ),
+      );
+      expect(e, equals('[Column:children=[RichText:(:Foo)],[Text:Bar]]'));
     });
   });
 
@@ -315,7 +358,7 @@ void main() {
 
     testWidgets('renders default value', (WidgetTester tester) async {
       final e = await explain(tester, HtmlWidget(html, key: helper.hwKey));
-      expect(e, equals('[RichText:(#FF0000FF+u:Foo)]'));
+      expect(e, equals('[RichText:(#FF123456+u:Foo)]'));
     });
 
     testWidgets('renders custom value', (WidgetTester tester) async {
@@ -331,11 +374,102 @@ void main() {
         tester,
         HtmlWidget(html, hyperlinkColor: null, key: helper.hwKey),
       );
-      expect(explained, equals('[RichText:(+u:Foo)]'));
+      expect(explained, equals('[RichText:(#FF123456+u:Foo)]'));
     });
   });
 
-  // TODO: onTapUrl
+  group('onTapUrl', () {
+    testWidgets('triggers callback (returns void)', (tester) async {
+      const href = 'returns-void';
+      final urls = <String>[];
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: urls.add,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(urls, equals(const [href]));
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('triggers callback (returns false)', (tester) async {
+      const href = 'returns-false';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) => false,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+
+    testWidgets('triggers callback (returns true)', (tester) async {
+      const href = 'returns-true';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) => true,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('triggers callback (async false)', (tester) async {
+      const href = 'returns-false';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) async => false,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+
+    testWidgets('triggers callback (async true)', (tester) async {
+      const href = 'returns-true';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) async => true,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('default handler', (WidgetTester tester) async {
+      const href = 'default';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+  });
 
   group('textStyle', () {
     final html = 'Foo';
@@ -357,4 +491,43 @@ void main() {
       expect(explained, equals('[RichText:(+i:Foo)]'));
     });
   });
+}
+
+class _OnTapUrlApp extends StatelessWidget {
+  final String href;
+  final dynamic Function(String)? onTapUrl;
+  final List? onTapCallbackResults;
+
+  const _OnTapUrlApp({
+    Key? key,
+    required this.href,
+    this.onTapCallbackResults,
+    this.onTapUrl,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext _) => MaterialApp(
+        home: Scaffold(
+          body: HtmlWidget(
+            '<a href="$href">Tap me</a>',
+            factoryBuilder: () => _OnTapUrlFactory(
+              onTapCallbackResults: onTapCallbackResults,
+            ),
+            onTapUrl: onTapUrl,
+          ),
+        ),
+      );
+}
+
+class _OnTapUrlFactory extends WidgetFactory {
+  final List? onTapCallbackResults;
+
+  _OnTapUrlFactory({this.onTapCallbackResults});
+
+  @override
+  Future<bool> onTapCallback(String url) async {
+    final result = await super.onTapCallback(url);
+    onTapCallbackResults?.addAll([url, result]);
+    return result;
+  }
 }
