@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:network_image_mock/network_image_mock.dart';
+import 'package:mocktail_image_network/mocktail_image_network.dart';
 
 import '_.dart';
 
@@ -9,9 +9,9 @@ void main() {
     final explained = await explain(tester, html);
     expect(
         explained,
-        equals('[RichText:[Stack:children='
-            '[Padding:(0,0,3,0),child=[Opacity:child=[RichText:(@8.3:Foo)]]],'
-            '[Positioned:(null,null,0.0,null),child=[RichText:(@8.3:Foo)]]'
+        equals('[RichText:'
+            '[Align:alignment=bottomCenter,child='
+            '[Padding:(3,0,0,0),child=[RichText:(@8.3:Foo)]]'
             ']@top]'));
   });
 
@@ -20,9 +20,9 @@ void main() {
     final explained = await explain(tester, html);
     expect(
         explained,
-        equals('[RichText:[Stack:children='
-            '[Padding:(3,0,0,0),child=[Opacity:child=[RichText:(@8.3:Foo)]]],'
-            '[Positioned:(0.0,null,null,null),child=[RichText:(@8.3:Foo)]]'
+        equals('[RichText:'
+            '[Align:alignment=topCenter,child='
+            '[Padding:(0,0,3,0),child=[RichText:(@8.3:Foo)]]'
             ']@bottom]'));
   });
 
@@ -49,14 +49,15 @@ void main() {
     final explained = await explain(tester, html);
     expect(explained, equals('[RichText:[RichText:(:Foo)]@middle]'));
   });
+
   testWidgets('renders sub text', (WidgetTester tester) async {
     final html = '<span style="vertical-align: sub">Foo</span>';
     final explained = await explain(tester, html);
     expect(
         explained,
-        equals('[RichText:[Stack:children='
-            '[Padding:(0,0,4,0),child=[Opacity:child=[RichText:(:Foo)]]],'
-            '[Positioned:(null,null,0.0,null),child=[RichText:(:Foo)]]'
+        equals('[RichText:'
+            '[Align:alignment=bottomCenter,child='
+            '[Padding:(4,0,0,0),child=[RichText:(:Foo)]]'
             ']@top]'));
   });
 
@@ -65,17 +66,20 @@ void main() {
     final explained = await explain(tester, html);
     expect(
         explained,
-        equals('[RichText:[Stack:children='
-            '[Padding:(4,0,0,0),child=[Opacity:child=[RichText:(:Foo)]]],'
-            '[Positioned:(0.0,null,null,null),child=[RichText:(:Foo)]]'
+        equals('[RichText:'
+            '[Align:alignment=topCenter,child='
+            '[Padding:(0,0,4,0),child=[RichText:(:Foo)]]'
             ']@bottom]'));
   });
 
   group('image', () {
     final imgSrc = 'http://domain.com/image.png';
-    final imgRendered = '[Image:image=NetworkImage("$imgSrc", scale: 1.0)]';
+    final imgRendered =
+        '[CssSizing:height≥0.0,height=auto,width≥0.0,width=auto,child='
+        '[Image:image=NetworkImage("$imgSrc", scale: 1.0)]'
+        ']';
     final imgExplain = (WidgetTester t, String html) =>
-        mockNetworkImagesFor(() => explain(t, html));
+        mockNetworkImages(() => explain(t, html));
 
     testWidgets('renders top image', (WidgetTester tester) async {
       final html = '<img src="$imgSrc" style="vertical-align: top" />';
@@ -101,9 +105,8 @@ void main() {
       expect(
           explained,
           equals('[RichText:(:$imgRendered'
-              '[Stack:children='
-              '[Padding:(3,0,0,0),child=[Opacity:child=[RichText:(@8.3:Foo)]]],'
-              '[Positioned:(0.0,null,null,null),child=[RichText:(@8.3:Foo)]]'
+              '[Align:alignment=topCenter,child='
+              '[Padding:(0,0,3,0),child=[RichText:(@8.3:Foo)]]'
               ']@bottom)]'));
     });
   });
@@ -118,6 +121,68 @@ void main() {
     final html = '<em><span style="vertical-align: top">Foo</span></em>';
     final explained = await explain(tester, html);
     expect(explained, equals('[RichText:[RichText:(+i:Foo)]@top]'));
+  });
+
+  group('isBlockElement', () {
+    group('renders top', () {
+      final html = '<div style="vertical-align: top">Foo</span>';
+
+      testWidgets('ltr', (WidgetTester tester) async {
+        final explained = await explain(tester, html);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=topLeft,child='
+                '[RichText:(:Foo)]]]'));
+      });
+
+      testWidgets('rtl', (WidgetTester tester) async {
+        final explained = await explain(tester, html, rtl: true);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=topRight,child='
+                '[RichText:dir=rtl,(:Foo)]]]'));
+      });
+    });
+
+    group('renders middle', () {
+      final html = '<div style="vertical-align: middle">Foo</span>';
+
+      testWidgets('ltr', (WidgetTester tester) async {
+        final explained = await explain(tester, html);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=centerLeft,child='
+                '[RichText:(:Foo)]]]'));
+      });
+
+      testWidgets('rtl', (WidgetTester tester) async {
+        final explained = await explain(tester, html, rtl: true);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=centerRight,child='
+                '[RichText:dir=rtl,(:Foo)]]]'));
+      });
+    });
+
+    group('renders bottom', () {
+      final html = '<div style="vertical-align: bottom">Foo</span>';
+
+      testWidgets('ltr', (WidgetTester tester) async {
+        final explained = await explain(tester, html);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=bottomLeft,child='
+                '[RichText:(:Foo)]]]'));
+      });
+
+      testWidgets('rtl', (WidgetTester tester) async {
+        final explained = await explain(tester, html, rtl: true);
+        expect(
+            explained,
+            equals('[CssBlock:child=[Align:alignment=bottomRight,child='
+                '[RichText:dir=rtl,(:Foo)]]]'));
+      });
+    });
   });
 
   group('error handling', () {
@@ -139,29 +204,10 @@ void main() {
       expect(explained, equals('[RichText:(:Foo)]'));
     });
 
-    testWidgets('#159: renders CODE of whitespaces inside', (tester) async {
-      final html = 'Foo <sup><code>\n  \n\n\n</code></sup>';
-      final explained = await explain(tester, html);
-      expect(explained, equals('[RichText:(:Foo)]'));
-    });
-
     testWidgets('#170: renders whitespace contents', (tester) async {
       final html = 'Foo <sub> </sub>';
       final explained = await explain(tester, html);
       expect(explained, equals('[RichText:(:Foo)]'));
-    });
-
-    testWidgets('#170: renders trailing whitespace contents', (tester) async {
-      final html = 'Foo <sub>bar </sub>';
-      final explained = await explain(tester, html);
-      expect(
-          explained,
-          equals('[RichText:(:'
-              'Foo '
-              '[Stack:children='
-              '[Padding:(0,0,3,0),child=[Opacity:child=[RichText:(@8.3:bar)]]],'
-              '[Positioned:(null,null,0.0,null),child=[RichText:(@8.3:bar)]]'
-              ']@top)]'));
     });
   });
 }
