@@ -1,4 +1,5 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 import 'package:flutter_widget_from_html_core/src/internal/tsh_widget.dart';
@@ -10,7 +11,7 @@ Future<String> explain(WidgetTester t, HtmlWidget hw) =>
 
 void main() {
   group('buildAsync', () {
-    final explain = (WidgetTester tester, String html, bool buildAsync) =>
+    final explain = (WidgetTester tester, String html, bool? buildAsync) =>
         tester.runAsync(() => helper.explain(tester, null,
             hw: HtmlWidget(
               html,
@@ -41,8 +42,8 @@ void main() {
     final explain = (
       WidgetTester tester,
       String html, {
-      AsyncWidgetBuilder<Widget> buildAsyncBuilder,
-      bool withData,
+      AsyncWidgetBuilder<Widget>? buildAsyncBuilder,
+      required bool withData,
     }) =>
         tester.runAsync(() => helper.explain(tester, null,
             buildFutureBuilderWithData: withData,
@@ -60,7 +61,8 @@ void main() {
         expect(explained, equals('[FutureBuilder:[RichText:(:$html)]]'));
       });
 
-      testWidgets('renders indicator', (WidgetTester tester) async {
+      testWidgets('renders CircularProgressIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.android;
         final html = 'Foo';
         final explained = await explain(tester, html, withData: false);
         expect(
@@ -70,13 +72,28 @@ void main() {
                 '[Padding:(8,8,8,8),child='
                 '[CircularProgressIndicator]'
                 ']]]'));
+        debugDefaultTargetPlatformOverride = null;
+      });
+
+      testWidgets('renders CupertinoActivityIndicator', (tester) async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+        final html = 'Foo';
+        final explained = await explain(tester, html, withData: false);
+        expect(
+            explained,
+            equals('[FutureBuilder:'
+                '[Center:child='
+                '[Padding:(8,8,8,8),child='
+                '[CupertinoActivityIndicator]'
+                ']]]'));
+        debugDefaultTargetPlatformOverride = null;
       });
     });
 
     group('custom', () {
       final buildAsyncBuilder =
           (BuildContext _, AsyncSnapshot<Widget> snapshot) =>
-              snapshot.hasData ? snapshot.data : Text('No data');
+              snapshot.data ?? Text('No data');
 
       testWidgets('renders data', (WidgetTester tester) async {
         final html = 'Foo';
@@ -107,12 +124,11 @@ void main() {
       WidgetTester tester,
       String html,
       bool enableCaching, {
-      Uri baseUrl,
-      bool buildAsync,
+      Uri? baseUrl,
+      bool? buildAsync,
       Color hyperlinkColor = const Color.fromRGBO(0, 0, 255, 1),
-      RebuildTriggers rebuildTriggers,
-      TextStyle textStyle,
-      bool unsupportedWebViewWorkaroundForIssue37 = false,
+      RebuildTriggers? rebuildTriggers,
+      TextStyle? textStyle,
       bool webView = false,
       bool webViewJs = true,
     }) =>
@@ -125,14 +141,12 @@ void main() {
               hyperlinkColor: hyperlinkColor,
               key: helper.hwKey,
               rebuildTriggers: rebuildTriggers,
-              textStyle: textStyle,
-              unsupportedWebViewWorkaroundForIssue37:
-                  unsupportedWebViewWorkaroundForIssue37,
+              textStyle: textStyle ?? const TextStyle(),
               webView: webView,
               webViewJs: webViewJs,
             ));
 
-    final _expect = (Widget built1, Widget built2, Matcher matcher) {
+    final _expect = (Widget? built1, Widget? built2, Matcher matcher) {
       final widget1 = (built1 as TshWidget).child;
       final widget2 = (built2 as TshWidget).child;
       expect(widget1 == widget2, matcher);
@@ -233,19 +247,6 @@ void main() {
       expect(explained2, equals('[RichText:(@20.0:Foo)]'));
     });
 
-    testWidgets('rebuild new workaroundForIssue37', (tester) async {
-      final html = 'Foo';
-
-      final explained1 = await explain(tester, html, true);
-      expect(explained1, equals('[RichText:(:Foo)]'));
-      final built1 = helper.buildCurrentState();
-
-      await explain(tester, html, true,
-          unsupportedWebViewWorkaroundForIssue37: true);
-      final built2 = helper.buildCurrentState();
-      _expect(built1, built2, isFalse);
-    });
-
     testWidgets('rebuild new webView', (tester) async {
       final html = 'Foo';
 
@@ -298,10 +299,10 @@ void main() {
       );
       expect(
           explained,
-          equals(
+          equals('[CssSizing:height≥0.0,height=auto,width≥0.0,width=auto,child='
               '[Image:image=CachedNetworkImageProvider("http://base.com/path/image.png", scale: 1.0),'
               'semanticLabel=image dot png'
-              ']'));
+              ']]'));
     });
   });
 
@@ -345,7 +346,37 @@ void main() {
           key: helper.hwKey,
         ),
       );
-      expect(explained, equals('[CssBlock:child=[Text:Bar]]'));
+      expect(explained, equals('[Text:Bar]'));
+    });
+  });
+
+  group('customWidgetBuilder (VIDEO)', () {
+    final CustomWidgetBuilder customWidgetBuilder =
+        (e) => e.localName == 'video' ? Text('Bar') : null;
+    final src = 'http://domain.com/video.mp4';
+    final html = 'Foo <video><source src="$src"></video>';
+
+    testWidgets('renders without value', (WidgetTester tester) async {
+      final explained =
+          await explain(tester, HtmlWidget(html, key: helper.hwKey));
+      expect(
+          explained,
+          equals('[Column:children='
+              '[RichText:(:Foo)],'
+              '[VideoPlayer:url=http://domain.com/video.mp4,aspectRatio=1.78]'
+              ']'));
+    });
+
+    testWidgets('renders with value', (WidgetTester tester) async {
+      final e = await explain(
+        tester,
+        HtmlWidget(
+          html,
+          customWidgetBuilder: customWidgetBuilder,
+          key: helper.hwKey,
+        ),
+      );
+      expect(e, equals('[Column:children=[RichText:(:Foo)],[Text:Bar]]'));
     });
   });
 
@@ -367,7 +398,98 @@ void main() {
     });
   });
 
-  // TODO: onTapUrl
+  group('onTapUrl', () {
+    testWidgets('triggers callback (returns void)', (tester) async {
+      const href = 'returns-void';
+      final urls = <String>[];
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: urls.add,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(urls, equals(const [href]));
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('triggers callback (returns false)', (tester) async {
+      const href = 'returns-false';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) => false,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+
+    testWidgets('triggers callback (returns true)', (tester) async {
+      const href = 'returns-true';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) => true,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('triggers callback (async false)', (tester) async {
+      const href = 'returns-false';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) async => false,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+
+    testWidgets('triggers callback (async true)', (tester) async {
+      const href = 'returns-true';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+        onTapUrl: (_) async => true,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, true]));
+    });
+
+    testWidgets('default handler', (WidgetTester tester) async {
+      const href = 'default';
+      final onTapCallbackResults = [];
+
+      await tester.pumpWidget(_OnTapUrlApp(
+        href: href,
+        onTapCallbackResults: onTapCallbackResults,
+      ));
+      await tester.pumpAndSettle();
+      expect(await helper.tapText(tester, 'Tap me'), equals(1));
+
+      expect(onTapCallbackResults, equals(const [href, false]));
+    });
+  });
 
   group('textStyle', () {
     final html = 'Foo';
@@ -397,10 +519,7 @@ void main() {
 
     testWidgets('renders default value', (WidgetTester tester) async {
       final e = await explain(tester, HtmlWidget(html, key: helper.hwKey));
-      expect(
-          e,
-          equals(
-              '[CssBlock:child=[GestureDetector:child=[Text:$webViewSrc]]]'));
+      expect(e, equals('[GestureDetector:child=[Text:$webViewSrc]]'));
     });
 
     testWidgets('renders true value', (WidgetTester tester) async {
@@ -413,10 +532,11 @@ void main() {
           ));
       expect(
           explained,
-          equals('[CssBlock:child=[WebView:'
+          equals('[WebView:'
               'url=$webViewSrc,'
               'aspectRatio=$webViewDefaultAspectRatio,'
-              'autoResize=true]]'));
+              'autoResize=true'
+              ']'));
     });
 
     testWidgets('renders false value', (WidgetTester tester) async {
@@ -427,44 +547,27 @@ void main() {
             key: helper.hwKey,
             webView: false,
           ));
-      expect(
-          explained,
-          equals(
-              '[CssBlock:child=[GestureDetector:child=[Text:$webViewSrc]]]'));
+      expect(explained, equals('[GestureDetector:child=[Text:$webViewSrc]]'));
     });
 
-    testWidgets('renders null value', (WidgetTester tester) async {
-      final explained = await explain(
-          tester,
-          HtmlWidget(
-            html,
-            key: helper.hwKey,
-            webView: null,
-          ));
-      expect(
-          explained,
-          equals(
-              '[CssBlock:child=[GestureDetector:child=[Text:$webViewSrc]]]'));
-    });
-
-    group('unsupportedWebViewWorkaroundForIssue37', () {
+    group('webViewDebuggingEnabled', () {
       testWidgets('renders true value', (WidgetTester tester) async {
         final explained = await explain(
             tester,
             HtmlWidget(
               html,
               key: helper.hwKey,
-              unsupportedWebViewWorkaroundForIssue37: true,
               webView: true,
+              webViewDebuggingEnabled: true,
             ));
         expect(
             explained,
-            equals('[CssBlock:child=[WebView:'
+            equals('[WebView:'
                 'url=$webViewSrc,'
                 'aspectRatio=$webViewDefaultAspectRatio,'
                 'autoResize=true,'
-                'unsupportedWorkaroundForIssue37=true'
-                ']]'));
+                'debuggingEnabled=true'
+                ']'));
       });
 
       testWidgets('renders false value', (WidgetTester tester) async {
@@ -473,34 +576,16 @@ void main() {
             HtmlWidget(
               html,
               key: helper.hwKey,
-              unsupportedWebViewWorkaroundForIssue37: false,
               webView: true,
+              webViewDebuggingEnabled: false,
             ));
         expect(
             explained,
-            equals('[CssBlock:child=[WebView:'
+            equals('[WebView:'
                 'url=$webViewSrc,'
                 'aspectRatio=$webViewDefaultAspectRatio,'
                 'autoResize=true'
-                ']]'));
-      });
-
-      testWidgets('renders null value', (WidgetTester tester) async {
-        final explained = await explain(
-            tester,
-            HtmlWidget(
-              html,
-              key: helper.hwKey,
-              unsupportedWebViewWorkaroundForIssue37: null,
-              webView: true,
-            ));
-        expect(
-            explained,
-            equals('[CssBlock:child=[WebView:'
-                'url=$webViewSrc,'
-                'aspectRatio=$webViewDefaultAspectRatio,'
-                'autoResize=true'
-                ']]'));
+                ']'));
       });
     });
 
@@ -516,11 +601,11 @@ void main() {
             ));
         expect(
             explained,
-            equals('[CssBlock:child=[WebView:'
+            equals('[WebView:'
                 'url=$webViewSrc,'
                 'aspectRatio=$webViewDefaultAspectRatio,'
                 'autoResize=true'
-                ']]'));
+                ']'));
       });
 
       testWidgets('renders false value', (WidgetTester tester) async {
@@ -534,11 +619,71 @@ void main() {
             ));
         expect(
             explained,
-            equals('[CssBlock:child=[WebView:'
+            equals('[WebView:'
                 'url=$webViewSrc,'
                 'aspectRatio=$webViewDefaultAspectRatio,'
                 'js=false'
-                ']]'));
+                ']'));
+      });
+    });
+
+    group('webViewMediaPlaybackAlwaysAllow', () {
+      testWidgets('renders true value', (WidgetTester tester) async {
+        final explained = await explain(
+            tester,
+            HtmlWidget(
+              html,
+              key: helper.hwKey,
+              webView: true,
+              webViewMediaPlaybackAlwaysAllow: true,
+            ));
+        expect(
+            explained,
+            equals('[WebView:'
+                'url=$webViewSrc,'
+                'aspectRatio=$webViewDefaultAspectRatio,'
+                'autoResize=true,'
+                'mediaPlaybackAlwaysAllow=true'
+                ']'));
+      });
+
+      testWidgets('renders false value', (WidgetTester tester) async {
+        final explained = await explain(
+            tester,
+            HtmlWidget(
+              html,
+              key: helper.hwKey,
+              webView: true,
+              webViewMediaPlaybackAlwaysAllow: false,
+            ));
+        expect(
+            explained,
+            equals('[WebView:'
+                'url=$webViewSrc,'
+                'aspectRatio=$webViewDefaultAspectRatio,'
+                'autoResize=true'
+                ']'));
+      });
+    });
+
+    group('webViewUserAgent', () {
+      testWidgets('renders string', (WidgetTester tester) async {
+        final explained = await explain(
+            tester,
+            HtmlWidget(
+              html,
+              key: helper.hwKey,
+              webView: true,
+              webViewUserAgent: 'Foo',
+            ));
+        expect(
+            explained,
+            equals('[WebView:'
+                'url=$webViewSrc,'
+                'aspectRatio=$webViewDefaultAspectRatio,'
+                'autoResize=true,'
+                'userAgent=Foo'
+                ']'));
       });
 
       testWidgets('renders null value', (WidgetTester tester) async {
@@ -548,16 +693,55 @@ void main() {
               html,
               key: helper.hwKey,
               webView: true,
-              webViewJs: null,
+              webViewUserAgent: null,
             ));
         expect(
             explained,
-            equals('[CssBlock:child=[WebView:'
+            equals('[WebView:'
                 'url=$webViewSrc,'
                 'aspectRatio=$webViewDefaultAspectRatio,'
-                'js=false'
-                ']]'));
+                'autoResize=true'
+                ']'));
       });
     });
   });
+}
+
+class _OnTapUrlApp extends StatelessWidget {
+  final String href;
+  final dynamic Function(String)? onTapUrl;
+  final List? onTapCallbackResults;
+
+  const _OnTapUrlApp({
+    Key? key,
+    required this.href,
+    this.onTapCallbackResults,
+    this.onTapUrl,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext _) => MaterialApp(
+        home: Scaffold(
+          body: HtmlWidget(
+            '<a href="$href">Tap me</a>',
+            factoryBuilder: () => _OnTapUrlFactory(
+              onTapCallbackResults: onTapCallbackResults,
+            ),
+            onTapUrl: onTapUrl,
+          ),
+        ),
+      );
+}
+
+class _OnTapUrlFactory extends WidgetFactory {
+  final List? onTapCallbackResults;
+
+  _OnTapUrlFactory({this.onTapCallbackResults});
+
+  @override
+  Future<bool> onTapCallback(String url) async {
+    final result = await super.onTapCallback(url);
+    onTapCallbackResults?.addAll([url, result]);
+    return result;
+  }
 }
